% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponse.R
\name{anydoseresponseC}
\alias{anydoseresponseC}
\title{Function to determine if there is a dose-response effect, for the normal distribution}
\usage{
anydoseresponseC(data)
}
\arguments{
\item{data}{Individual data with columns: dose, response, litter}
}
\value{
list containing Bayes factor and decision.
}
\description{
This function tests for any dose-response effect using Bayes factor.
It fits a null model and a saturated model and compares these two models using model posterior probabilities obtained via bridge sampling.
Currently only based on the normal distribution.
}
