% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponseLN.R
\name{anydoseresponseLNI}
\alias{anydoseresponseLNI}
\alias{anydoseresponseLND}
\title{Function to determine if there is a dose-response effect}
\usage{
anydoseresponseLNI(dose.a, mean.a, sd.a, n.a)

anydoseresponseLND(dose.a, mean.a, sd.a, n.a)
}
\arguments{
\item{dose.a}{dose levels}

\item{mean.a}{mean response}

\item{sd.a}{standard deviation}

\item{n.a}{number of observations per dose level}
}
\value{
text
}
\description{
Function to determine if there is a dose-response effect
}
