% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMABMDR-package.R, R/packageInfo.R
\docType{package}
\name{BMABMDR-package}
\alias{BMABMDR-package}
\alias{BMABMDR}
\title{The 'BMABMDR' package.}
\description{
A DESCRIPTION OF THE PACKAGE

Here comes the information of the package
}
\details{
Some more information
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.3. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cecilekremer/BMABMDR}
  \item Report bugs at \url{https://github.com/cecilekremer/BMABMDR/issues}
}

}
\author{
\strong{Maintainer}: Cécile Kremer \email{cecile.kremer@uhasselt.be} (\href{https://orcid.org/0000-0002-2081-7662}{ORCID})

Authors:
\itemize{
  \item Cécile Kremer \email{cecile.kremer@uhasselt.be} (\href{https://orcid.org/0000-0002-2081-7662}{ORCID})
  \item Oluwafemi Olusoji
  \item Marc Aerts
  \item Ziv Shkedy
  \item José Cortinas Abrahantes
}

Other contributors:
\itemize{
  \item Machteld Varewyck [contributor]
  \item Wouter Verlinden [contributor]
}

}
