% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputC.R, R/outputQ.R
\name{BMDLU}
\alias{BMDLU}
\title{function to extract the BMDL,BMD and BMDU}
\usage{
BMDLU(mod.obj, pvec = c(0.05, 0.5, 0.95), type = c("continuous", "quantal"))

BMDLU(mod.obj, pvec = c(0.05, 0.5, 0.95), type = c("continuous", "quantal"))
}
\arguments{
\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDLs, BMDs and BMDUs per model

datatframe containing BMDLs, BMDs and BMDUs per model
}
\description{
function to extract the BMDL,BMD and BMDU

function to extract the BMDL,BMD and BMDU
}
