% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDmixture_extract}
\alias{BMDmixture_extract}
\alias{BMDQmixture_extract}
\title{Function to extract BMD mixture}
\usage{
BMDmixture_extract(mod.obj, weight_type, conv = FALSE)

BMDQmixture_extract(mod.obj, weight_type, conv = FALSE)
}
\arguments{
\item{weight_type}{character indicating whether to use Laplace ('LP') or Bridge sampling ('BS') weights}

\item{conv}{logical to indicate if mixture should be only based on converged
models or all the models regardless of the convergence status}

\item{mod_obj}{BMDBMA model object}
}
\value{
dataframe containing Mixture per model
}
\description{
Function to extract BMD mixture
}
