% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDQmixture_extract}
\alias{BMDQmixture_extract}
\title{Function for internal use}
\usage{
BMDQmixture_extract(mod.obj, conv = FALSE)
}
\arguments{
\item{mod.obj}{value}

\item{conv}{value}
}
\value{
.
}
\description{
Function for internal use
}
