% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{convergence_deci}
\alias{convergence_deci}
\title{Function for internal use}
\usage{
convergence_deci(
  model_stan,
  nrchains = 3,
  pars = c("par[2]", letters[c(1, 3, 4)], "invsigma2", "lp__")
)
}
\arguments{
\item{model_stan}{stan model}

\item{pars}{parameters}
}
\value{
.
}
\description{
Function for internal use
}
