% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{density_BMD}
\alias{density_BMD}
\title{Function for internal use}
\usage{
density_BMD(
  BMD_DF,
  BMD.true = NULL,
  to_plot = c("BMD", "BMDL", "BMDU"),
  xlab = expression(log[10](BMD)),
  title = "MLE"
)
}
\arguments{
\item{BMD_DF}{dataframe containing BMD, BMDL and BMDU}

\item{BMD.true}{true value for BMD. Only necessary if such exist}

\item{to_plot}{which quantity to plot}

\item{xlab}{x-axis label}

\item{title}{title}
}
\value{
.
}
\description{
Function for internal use
}
