% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{DR_scatter}
\alias{DR_scatter}
\title{Function for internal use}
\usage{
DR_scatter(
  DR_df = NULL,
  DRM_df = NULL,
  x_DR_df,
  y_DR_df,
  x_DRM_df,
  y_DRM_df,
  col_DRM = brewer.pal(3, "Set1")[1]
)
}
\arguments{
\item{DR_df}{dataframe containing individual-level dose response data}

\item{DRM_df}{dataframe for dose-response model}

\item{x_DR_df}{name of the x varaibel in the individual-level dose response data.
Typically dose.}

\item{y_DR_df}{name of the y variable in the individual-level dose response data.}

\item{x_DRM_df}{name of the x variable in the dose response curve data. Typically dose}

\item{y_DRM_df}{name of the y variable in the individual-level dose response data.}

\item{col_DRM}{color scheme}
}
\value{
.
}
\description{
Function for internal use
}
