% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM.R
\name{DRM.E4_Q}
\alias{DRM.E4_Q}
\alias{DRM.IE4_Q}
\alias{DRM.H4_Q}
\alias{DRM.LN4_Q}
\alias{DRM.G4_Q}
\alias{DRM.QE4_Q}
\alias{DRM.P4_Q}
\alias{DRM.L4_Q}
\title{The dose response models}
\usage{
DRM.E4_Q(par, x, q)

DRM.IE4_Q(par, x, q)

DRM.H4_Q(par, x, q)

DRM.LN4_Q(par, x, q)

DRM.G4_Q(par, x, q)

DRM.QE4_Q(par, x, q)

DRM.P4_Q(par, x, q)

DRM.L4_Q(par, x, q)
}
\arguments{
\item{par}{parameters in the order background, BMD, log(d)}

\item{x}{unique ordered dose levels}

\item{q}{specified BMR}
}
\value{
Vector containing the expected response at each dose level
}
\description{
These are the dose response models used internally in the BMA functions.
}
