% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{DRM}
\alias{DRM}
\title{Function for internal use}
\usage{
DRM(Model, type = c("increasing", "decreasing"))
}
\arguments{
\item{Model}{which model}

\item{type}{increasing or decreasing}
}
\value{
.
}
\description{
Function for internal use
}
