% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM.R
\name{DRMQ}
\alias{DRMQ}
\title{Function to select a dose response model based on model name}
\usage{
DRMQ(Model)
}
\arguments{
\item{Model}{a character for the model name.It can be either of E4_Q, IE4_Q, H4_Q, LN4_Q, G4_Q, QE4_Q,
P4_Q or L4_Q}
}
\value{
a corresponding DRM function
}
\description{
Function to select a dose response model based on model name
}
