% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_pert.R
\name{fit_pert}
\alias{fit_pert}
\title{Function to fit PERT distribution}
\usage{
fit_pert(mod.obj, param = "BMD", method = "MLE")
}
\arguments{
\item{mod.obj}{an object of class BMADR}

\item{param}{which parameter's posterior the PERT should be fitted to (BMD, bkg, maxy)}

\item{method}{estimation method to use (MLE, JKL)}
}
\value{
Vector containing the parameters of a PERT distribution
}
\description{
This function can be used to estimate the shape parameter of a PERT distribution by fitting to a posterior
}
