% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{flat2}
\alias{flat2}
\title{Function to check if the mean for the Dose-Response data is approaching or at its asymptote
for internal use}
\usage{
flat2(dose, response, h = 1/1000, B = 200)
}
\arguments{
\item{dose}{value}

\item{response}{value}

\item{h}{step size for the derivative}

\item{B}{number of bootstraps}
}
\value{
.
}
\description{
Function to check if the mean for the Dose-Response data is approaching or at its asymptote
for internal use
}
