% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Laplace.R
\name{full.laplace_MA}
\alias{full.laplace_MA}
\title{Perform model averaging using Full Laplace method}
\usage{
full.laplace_MA(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95),
  plot = FALSE
)
}
\arguments{
\item{data.N}{the input data as returned by function PREP_DATA_N}

\item{data.LN}{the input data as returned by function PREP_DATA_LN}

\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws, default 30000}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}

\item{plot}{logical indicating whether a simple plot of model fits should be shown}

\item{priordist}{which prior distribution should be used (currently only "PERT" is implemented)}

\item{prior.BMD}{logical indicating whether an informative prior for the BMD is used (currently not implemented)}
}
\value{
\enumerate{
\item E4_N parameter estimates from the exponential model
\item IE4_N parameter estimates from the inverse-exponential model
\item H4_N parameter estimates from the Hill model
\item LN4_N parameter estimates from the lognormal
\item G4_N parameter estimates from the gamma model
\item QE4_N parameter estimates from the quadratic-exponential model
\item P4_N parameter estimates from the probit model
\item L4_N parameter estimates from the logit model
\item E4_LN parameter estimates from the exponential model
\item IE4_LN parameter estimates from the inverse-exponential model
\item H4_LN parameter estimates from the Hill model
\item LN4_LN parameter estimates from the lognormal
\item G4_LN parameter estimates from the gamma model
\item QE4_LN parameter estimates from the quadratic-exponential model
\item P4_LN parameter estimates from the probit model
\item L4_LN parameter estimates from the logit model
\item MA_laplace Laplace approximation model averaged BMD estimates using all the models
\item MA_lp_conv Laplace approximation model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item llN vector of model likelihoods for normal distribution
\item llLN vector of model likelihoods for lognormal distribution
\item bf Bayes factor comparing the best model against saturated ANOVA model
}
}
\description{
Using Laplace approximation, we compute the parameters and weights of each model.
}
\details{
This method assumed data for continuous endpoints.

More detailed descriprion
}
\examples{
 # we use the first 5 rows because those are observations from subjects belonging to the same group.
 data("immunotoxicityData.rda")  #load the immunotoxicity data
 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]),
                       sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors
 data_LN <- PREP_DATA_LN(data = as.data.frame(immunotoxicityData[1:5,]),
                         sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors

 pvec <- c(0.05, 0.5, 0.95)
 prior.weights <- rep(1, 16); ndr <- 30000
 FLBMD <- full.laplace_MA(data_N,data_LN,prior.weights,
                       ndraws=ndr,seed=123,pvec=pvec)

}
