% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Laplace.R
\name{full.laplace_MA_Q_Cov}
\alias{full.laplace_MA_Q_Cov}
\title{Perform model averaging using Full Laplace method}
\usage{
full.laplace_MA_Q_Cov(
  data,
  sumstats = TRUE,
  q = q,
  prior.d = prior.d,
  extended = TRUE,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data}{the summary data, with columns: dose, number of adverse events, n, covariate}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{q}{specified BMR}

\item{prior.d}{prior distribution for parameter d, should be either N11 (default) or EPA}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)#}

\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws, default 30000}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}
}
\value{
\enumerate{
\item E4_Q parameter estimates from the exponential model
\item IE4_Q parameter estimates from the inverse-exponential model
\item H4_Q parameter estimates from the Hill model
\item LN4_Q parameter estimates from the lognormal
\item G4_Q parameter estimates from the gamma model
\item QE4_Q parameter estimates from the quadratic-exponential model
\item P4_Q parameter estimates from the probit model
\item L4_Q parameter estimates from the logit model
\item MA  Model averaged BMD credible interval
\item weights Model weights used in the averaging
}
}
\description{
Using Laplace approximation, we compute the parameters and weights of each model.
}
