% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LaplaceC.R
\name{full.laplace_MAc}
\alias{full.laplace_MAc}
\title{function to perform model-averaging using Laplace approximation}
\usage{
full.laplace_MAc(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data.N}{list containing data values for the normal models}

\item{data.LN}{list containing data values for the lognormal models}

\item{prior.weights}{model weights determining if a model is to be included in the model averaging or not.
1 implies model is included, 0 otherwise. Defaults to rep(1,16).}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{seed}{random seed for reproducibility. Defaults to 123}

\item{pvec}{probability vector to compute credible interval for the BMD. Defaults to c(0.05,0.5,0.95).}

\item{plot}{logical variable to determine if the results should be plotted. Defaults to FALSE}
}
\value{
list containing the following results
\enumerate{
\item E4_N parameter estimates from the exponential model with normal distribution assumed
\item IE4_N parameter estimates from the inverse-exponential model with normal distribution assumed
\item H4_N parameter estimates from the Hill model with normal distribution assumed
\item LN4_N parameter estimates from the lognormal model with normal distribution assumed
\item G4_N parameter estimates from the gamma model with normal distribution assumed
\item QE4_N parameter estimates from the quadratic-exponential model with normal distribution assumed
\item P4_N parameter estimates from the probit model with normal distribution assumed
\item L4_N parameter estimates from the logit model with normal distribution assumed
\item E4_LN parameter estimates from the exponential model with lognormal distribution assumed
\item IE4_LN parameter estimates from the inverse exponential model with lognormal distribution assumed
\item H4_LN parameter estimates from the Hill model with lognormal distribution assumed
\item LN4_LN parameter estimates from the lognormal model with lognormal distribution assumed
\item G4_LN parameter estimates from the gamma model with lognormal distribution assumed
\item QE4_LN parameter estimates from the quadratic exponential model with lognormal distribution assumed
\item P4_LN parameter estimates from the probit model with lognormal distribution assumed
\item L4_LN parameter estimates from the logit model with lognormal distribution assumed
\item MA_laplace Laplace approximation model averaged BMD estimates using all the models
\item MA_lp_conv Laplace approximation model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item convergence vector indicating model convergence or not. 1 = converged, 0 otherwise.
\item llN vector of model likelihoods for normal distribution
\item llLN vector of model likelihoods for lognonormal distribution
\item bf Bayes factor comparing the best model against saturated ANOVA model
}
}
\description{
function to perform model-averaging using Laplace approximation
}
\examples{

}
