% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Laplace.R
\name{full.laplaceQ_MA}
\alias{full.laplaceQ_MA}
\title{Perform model averaging using Full Laplace method}
\usage{
full.laplaceQ_MA(
  data.Q,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data.Q}{the input data as returned by function PREP_DATA_QE}

\item{prior.weights}{a vector specifying which of the 8 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws, default 30000}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}
}
\value{
\enumerate{
\item E4_Q parameter estimates from the exponential model
\item IE4_Q parameter estimates from the inverse-exponential model
\item H4_Q parameter estimates from the Hill model
\item LN4_Q parameter estimates from the lognormal
\item G4_Q parameter estimates from the gamma model
\item QE4_Q parameter estimates from the quadratic-exponential model
\item P4_Q parameter estimates from the probit model
\item L4_Q parameter estimates from the logit model
\item MA  Model averaged BMD credible interval
\item weights Model weights used in the averaging
\item bf Bayes factor comparing the best model against saturated ANOVA model
}
}
\description{
Using Laplace approximation, we compute the parameters and weights of each model.
}
