% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{fun.alpha}
\alias{fun.alpha}
\alias{fun.beta}
\title{Functions to get parameters for PERT prior (internal use)}
\usage{
fun.alpha(a, b, c, g)

fun.beta(a, b, c, g)
}
\arguments{
\item{a}{minimum}

\item{b}{most likely value}

\item{c}{maximum}

\item{g}{shape}
}
\value{
.
}
\description{
Functions to get parameters for PERT prior (internal use)
}
