% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_pert.R
\name{fun.quantile.dist}
\alias{fun.quantile.dist}
\title{Function to get 1000 quantiles}
\usage{
fun.quantile.dist(x)
}
\arguments{
\item{x}{vector of observations}
}
\value{
vector of quantiles
}
\description{
Function to get 1000 quantiles
}
