% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cov_selection.R
\name{fun.w}
\alias{fun.w}
\alias{fun.w.QE4}
\alias{fun.wQ}
\alias{fun.wQ.QE4}
\title{Function to calculate weights for Laplace including a covariate effect (used internally)}
\usage{
fun.w(
  optMod,
  lld,
  min.ll,
  nlevels,
  dataMod,
  covar = c("a_sigma2", "BMD_d", "all", "none")
)

fun.w.QE4(
  optMod,
  lld,
  min.ll,
  nlevels,
  dataMod,
  covar = c("a_sigma2", "BMD_d", "all", "none")
)

fun.wQ(
  optMod,
  lld,
  min.ll,
  nlevels,
  dataMod,
  covar = c("background", "BMD_d", "all", "none")
)

fun.wQ.QE4(
  optMod,
  lld,
  min.ll,
  nlevels,
  dataMod,
  covar = c("background", "BMD_d", "all", "none")
)
}
\arguments{
\item{optMod}{model}

\item{lld}{loglikelihood}

\item{min.ll}{minimum loglikelihood among submodels}

\item{nlevels}{number of covariate levels}

\item{dataMod}{data for optMod}

\item{covar}{which parameters depend on covariate}
}
\value{
.
}
\description{
This function computes Laplace weights for a submodel
}
\examples{
\dontrun{
fun.w(optMod_asigma2, lld = ll_asigma2, min.ll = minll, nlevels = data_asigma2$data$nlevels, dataMod = data_asigma2, covar = 'a_sigma2')
}

}
