% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cov_selection.R
\name{fun_cov_selection}
\alias{fun_cov_selection}
\alias{fun_cov_selectionQ}
\title{Function to perform model selection when including a covariate effect (used internally)}
\usage{
fun_cov_selection(
  model,
  model_name,
  model.none,
  loglik,
  data_asigma2,
  data_dBMD,
  data_all,
  data_none,
  prior.weightsCov = rep(1, 4),
  pvec,
  ndraws,
  td,
  seed
)

fun_cov_selectionQ(
  model,
  model_name,
  model.none,
  loglik,
  data_bkg,
  data_dBMD,
  data_all,
  data_none,
  prior.weightsCov = rep(1, 4),
  pvec,
  ndraws,
  td,
  seed
)
}
\arguments{
\item{model}{stan model including covariates}

\item{model_name}{model name}

\item{model.none}{stan model without covariates}

\item{loglik}{function to compute log-likelihood}

\item{data_asigma2}{data when covariate is a_sigma2}

\item{data_dBMD}{data when covariate is BMD_d}

\item{data_all}{data when covariate is all}

\item{data_none}{data when covariate is none}

\item{prior.weightsCov}{vector of prior weights for the 4 submodels (defaults to rep(1,4), equal weight)}

\item{pvec}{vector of probabilities for BMD credible interval}

\item{ndraws}{ndraws}

\item{td}{truncation for d in QE4 model}

\item{seed}{random seed for reproducibility}
}
\value{
.
}
\description{
This function selects the best-fitting submodel
}
