% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fit_pert.R
\name{fun_loglik}
\alias{fun_loglik}
\title{Function for loglikelihood of PERT distribution}
\usage{
fun_loglik(m, s, lb, ub)
}
\arguments{
\item{m}{mode}

\item{s}{shape}

\item{lb}{minimum}

\item{ub}{maximum}
}
\value{
loglikelihood
}
\description{
Function for loglikelihood of PERT distribution
}
