% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimQ.R
\name{fun_optimQ}
\alias{fun_optimQ}
\title{Function for internal use}
\usage{
fun_optimQ(mod, data, stv, ndraws = ndraws, seed = seed, pvec)
}
\arguments{
\item{mod}{stan model}

\item{data}{input data}

\item{stv}{start values}

\item{ndraws}{ndraws}

\item{seed}{seed}

\item{pvec}{pvec}
}
\value{
.
}
\description{
Function for internal use
}
