% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mcmcC.R
\name{fun_samplingC}
\alias{fun_samplingC}
\title{function to perform MCMC sampling for a given dose-response model.}
\usage{
fun_samplingC(
  mod,
  data,
  stv,
  ndraws = ndraws,
  nrchains = nrchains,
  nriterations = nriterations,
  warmup = warmup,
  delta = delta,
  treedepth = treedepth,
  seed = seed,
  pvec
)
}
\arguments{
\item{mod}{stan model}

\item{data}{list containing data values to be passed to the stan model file}

\item{stv}{list of starting values}

\item{ndraws}{number of draws to be made from the posterior distribution}

\item{nrchains}{number of MCMC chains}

\item{nriterations}{number of MCMC iterations}

\item{warmup}{number of MCMC iterations for warmup}

\item{delta}{adapt_delta value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.}

\item{treedepth}{tree_depth value for the HMC in stan. See \code{\link[rstan]{sampling}} for more}

\item{seed}{random seed for reproducibility}

\item{pvec}{probability vector to compute credible interval for the BMD}
}
\value{
a stan results object
}
\description{
function to perform MCMC sampling for a given dose-response model.
}
\examples{

}
