% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mcmcQ.R
\name{fun_samplingQ}
\alias{fun_samplingQ}
\title{Function for internal use}
\usage{
fun_samplingQ(
  mod,
  data,
  stv,
  ndraws = ndraws,
  nrchains = nrchains,
  nriterations = nriterations,
  warmup = warmup,
  delta = delta,
  treedepth = treedepth,
  seed = seed,
  pvec
)
}
\arguments{
\item{mod}{stan model}

\item{data}{input data}

\item{stv}{start values}

\item{ndraws}{ndraws}

\item{nrchains}{nrchains}

\item{nriterations}{nriterations}

\item{warmup}{burnin}

\item{delta}{delta}

\item{treedepth}{treedepth}

\item{seed}{seed}

\item{pvec}{pvec}
}
\value{
.
}
\description{
Function for internal use
}
