% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{get_models}
\alias{get_models}
\title{Obtain a list of all available models}
\usage{
get_models(type = c("continuous", "quantal"))
}
\arguments{
\item{type}{one of 'continuous' or 'quantal'}
}
\value{
list of all available models
}
\description{
Obtain a list of all available models
}
