% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBMD.R
\name{getBMD}
\alias{getBMD}
\alias{getBMD_Q}
\title{Function for internal use}
\usage{
getBMD(pars, pvec, covar, model_name)

getBMD_Q(pars, pvec, covar, model_name)
}
\arguments{
\item{pars}{value}

\item{pvec}{value}

\item{covar}{value}

\item{model_name}{value}
}
\value{
.
}
\description{
Function for internal use
}
