% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunotoxicityData}
\alias{immunotoxicityData}
\title{immunotoxicity dataset. Effects on Th17 cells}
\format{
A data frame with 20 rows and 5 variables:
\describe{
\item{Dose}{dosage}
\item{Mean}{average number of cells per dose}
\item{SD}{standard deviation per dose}
\item{n}{number of observations in the dose group}
\item{group}{animal group (sex)}
}
}
\source{
Luo SM, Li Y, Li YP, Zhu QX, Jiang JH, Wu CH and Shen T, 2016. Gestational and lactational exposure
to low-dose bisphenol A increases Th17 cells in mice offspring. Environmental Toxicology and
Pharmacology, 47, 149–158. doi: 10.1016/j.etap.2016.09.017. RefID: 4679.
}
\usage{
immunotoxicityData
}
\description{
Th17 cell frequency in the spleen in offspring mice (\%) was analysed. In the paper, the data are
represented in a graph and the actual numbers were provided to EFSA by T. Shen. Based on the
provided information, EFSA calculated the standard deviation (SD).
}
\keyword{datasets}
