% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.BMADR2}
\alias{is.BMADR2}
\title{function to check if object is of class BMADR and its subcalsses}
\usage{
is.BMADR2(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class stanfitOptim. FALSE otherwise
}
\description{
function to check if object is of class BMADR and its subcalsses
}
\examples{
 x <- c(1, 5, 4)
 is.BMADR2(x)
}
