% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.stanfit}
\alias{is.stanfit}
\title{function to check if object is of class stanfit}
\usage{
is.stanfit(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class stanfit. FALSE otherwise
}
\description{
function to check if object is of class stanfit
}
\examples{
 x <- c(1, 5, 4)
 is.stanfit(x)
}
