% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.stanfitOptim}
\alias{is.stanfitOptim}
\title{function to check if object is of class stanfit optim}
\usage{
is.stanfitOptim(x)
}
\arguments{
\item{x}{any R object}
}
\value{
TRUE if object is of class stanfitOptim. FALSE otherwise
}
\description{
function to check if object is of class stanfit optim
}
\examples{
 x <- c(1, 5, 4)
 is.stanfit(x)
}
