% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LearningMemory}
\alias{LearningMemory}
\title{Learning and memory during the growth phase/young age exposure period}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{Dose}{dosage}
\item{Mean}{average per dose}
\item{SEM}{standard error of the mean per dose}
\item{n}{number of observations in the dose group}
}
}
\source{
Chen Z, Li T, Zhang L, Wang H and Hu F, 2018. Bisphenol A exposure remodels cognition of male rats
attributable to excitatory alterations in the hippocampus and visual cortex. Toxicology, 410, 132–
doi: 10.1016/j.tox.2018.10.002. RefID: 11734.
}
\usage{
LearningMemory
}
\description{
Learning and memory during the growth phase/young age exposure period
}
\keyword{datasets}
