% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_NI}
\alias{llfE4_NI}
\alias{llfIE4_NI}
\alias{llfH4_NI}
\alias{llfLN4_NI}
\alias{llfG4_NI}
\alias{llfQE4_NI}
\alias{llfP4_NI}
\alias{llfL4_NI}
\alias{llfE4_LNI}
\alias{llfIE4_LNI}
\alias{llfH4_LNI}
\alias{llfLN4_LNI}
\alias{llfG4_LNI}
\alias{llfQE4_LNI}
\alias{llfP4_LNI}
\alias{llfL4_LNI}
\alias{llfE4_ND}
\alias{llfIE4_ND}
\alias{llfH4_ND}
\alias{llfLN4_ND}
\alias{llfG4_ND}
\alias{llfQE4_ND}
\alias{llfP4_ND}
\alias{llfL4_ND}
\alias{llfE4_LND}
\alias{llfIE4_LND}
\alias{llfH4_LND}
\alias{llfLN4_LND}
\alias{llfG4_LND}
\alias{llfQE4_LND}
\alias{llfP4_LND}
\alias{llfL4_LND}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfH4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfG4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfP4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfL4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfE4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfH4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfG4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfP4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfL4_LNI(x, nvec, dvec, mvec, s2vec, qval)

llfE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfH4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfG4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfP4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfL4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfE4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfH4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfG4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfP4_LND(x, nvec, dvec, mvec, s2vec, qval)

llfL4_LND(x, nvec, dvec, mvec, s2vec, qval)
}
\arguments{
\item{x}{parameter values}

\item{nvec}{number of observations}

\item{dvec}{dose levels}

\item{mvec}{response}

\item{s2vec}{variance}

\item{qval}{BMR}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
