% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_NI_Cov}
\alias{llfE4_NI_Cov}
\alias{llfIE4_NI_Cov}
\alias{llfH4_NI_Cov}
\alias{llfLN4_NI_Cov}
\alias{llfG4_NI_Cov}
\alias{llfQE4_NI_Cov}
\alias{llfP4_NI_Cov}
\alias{llfL4_NI_Cov}
\alias{llfE4_LNI_Cov}
\alias{llfIE4_LNI_Cov}
\alias{llfH4_LNI_Cov}
\alias{llfLN4_LNI_Cov}
\alias{llfG4_LNI_Cov}
\alias{llfQE4_LNI_Cov}
\alias{llfP4_LNI_Cov}
\alias{llfL4_LNI_Cov}
\alias{llfE4_ND_Cov}
\alias{llfIE4_ND_Cov}
\alias{llfH4_ND_Cov}
\alias{llfLN4_ND_Cov}
\alias{llfG4_ND_Cov}
\alias{llfQE4_ND_Cov}
\alias{llfP4_ND_Cov}
\alias{llfL4_ND_Cov}
\alias{llfE4_LND_Cov}
\alias{llfIE4_LND_Cov}
\alias{llfH4_LND_Cov}
\alias{llfLN4_LND_Cov}
\alias{llfG4_LND_Cov}
\alias{llfQE4_LND_Cov}
\alias{llfP4_LND_Cov}
\alias{llfL4_LND_Cov}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfIE4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfH4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfLN4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfG4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfQE4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfP4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfL4_NI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfE4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfIE4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfH4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfLN4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfG4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfQE4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfP4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfL4_LNI_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfE4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfIE4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfH4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfLN4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfG4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfQE4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfP4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfL4_ND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfE4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfIE4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfH4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfLN4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfG4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfQE4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfP4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfL4_LND_Cov(
  pars,
  x,
  n,
  m,
  s2,
  qval,
  shift,
  covar = c("a_sigma2", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)
}
\arguments{
\item{pars}{parameter values}

\item{x}{vector containing the unique ordered dose levels}

\item{n}{vector containing the number of observations at each dose level}

\item{m}{vector containing the mean response at each dose level}

\item{s2}{vector containing the variance at each dose level}

\item{qval}{BMR}

\item{shift}{value of the shift for negative geometric means}

\item{covar}{which parameter includes a covariate effect}

\item{nlevels}{number of covariate levels}

\item{trt_ind}{matrix indicating which reponse corresponds to which covariate level}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
\examples{

}
