% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_NIc}
\alias{llfE4_NIc}
\alias{llfIE4_NIc}
\alias{llfH4_NIc}
\alias{llfLN4_NIc}
\alias{llfG4_NIc}
\alias{llfQE4_NIc}
\alias{llfP4_NIc}
\alias{llfL4_NIc}
\alias{llfE4_LNIc}
\alias{llfIE4_LNIc}
\alias{llfH4_LNIc}
\alias{llfLN4_LNIc}
\alias{llfG4_LNIc}
\alias{llfQE4_LNIc}
\alias{llfP4_LNIc}
\alias{llfL4_LNIc}
\alias{llfSM_LNc}
\alias{llfSM_Nc}
\alias{llfE4_NDc}
\alias{llfIE4_NDc}
\alias{llfH4_NDc}
\alias{llfLN4_NDc}
\alias{llfG4_NDc}
\alias{llfQE4_NDc}
\alias{llfP4_NDc}
\alias{llfL4_NDc}
\alias{llfE4_LNDc}
\alias{llfIE4_LNDc}
\alias{llfH4_LNDc}
\alias{llfLN4_LNDc}
\alias{llfG4_LNDc}
\alias{llfQE4_LNDc}
\alias{llfP4_LNDc}
\alias{llfL4_LNDc}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE4_NIc(x, d, n, nij, y, qval)

llfIE4_NIc(x, d, n, nij, y, qval)

llfH4_NIc(x, d, n, nij, y, qval)

llfLN4_NIc(x, d, n, nij, y, qval)

llfG4_NIc(x, d, n, nij, y, qval)

llfQE4_NIc(x, d, n, nij, y, qval)

llfP4_NIc(x, d, n, nij, y, qval)

llfL4_NIc(x, d, n, nij, y, qval)

llfE4_LNIc(x, d, n, nij, y, qval, shift)

llfIE4_LNIc(x, d, n, nij, y, qval, shift)

llfH4_LNIc(x, d, n, nij, y, qval, shift)

llfLN4_LNIc(x, d, n, nij, y, qval, shift)

llfG4_LNIc(x, d, n, nij, y, qval, shift)

llfQE4_LNIc(x, d, n, nij, y, qval, shift)

llfP4_LNIc(x, d, n, nij, y, qval, shift)

llfL4_LNIc(x, d, n, nij, y, qval, shift)

llfSM_LNc(x, d, n, nij, y, qval, shift)

llfSM_Nc(x, d, n, nij, y, qval)

llfE4_NDc(x, d, n, nij, y, qval)

llfIE4_NDc(x, d, n, nij, y, qval)

llfH4_NDc(x, d, n, nij, y, qval)

llfLN4_NDc(x, d, n, nij, y, qval)

llfG4_NDc(x, d, n, nij, y, qval)

llfQE4_NDc(x, d, n, nij, y, qval)

llfP4_NDc(x, d, n, nij, y, qval)

llfL4_NDc(x, d, n, nij, y, qval)

llfE4_LNDc(x, d, n, nij, y, qval, shift)

llfIE4_LNDc(x, d, n, nij, y, qval, shift)

llfH4_LNDc(x, d, n, nij, y, qval, shift)

llfLN4_LNDc(x, d, n, nij, y, qval, shift)

llfG4_LNDc(x, d, n, nij, y, qval, shift)

llfQE4_LNDc(x, d, n, nij, y, qval, shift)

llfP4_LNDc(x, d, n, nij, y, qval, shift)

llfL4_LNDc(x, d, n, nij, y, qval, shift)
}
\arguments{
\item{x}{parameter values}

\item{d}{dose levels}

\item{n}{observations per dose level}

\item{nij}{observations per dose x litter combination}

\item{y}{observed responses}

\item{qval}{BMR}

\item{shift}{value of the shift for negative geometric means}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
\examples{

}
