% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_Q}
\alias{llfE4_Q}
\alias{llfIE4_Q}
\alias{llfH4_Q}
\alias{llfLN4_Q}
\alias{llfG4_Q}
\alias{llfQE4_Q}
\alias{llfP4_Q}
\alias{llfL4_Q}
\alias{llfSM_Q}
\alias{llfH0_Q}
\title{Loglikelihood functions for the different dose response models, for quantal data}
\usage{
llfE4_Q(x, nvec, dvec, yvec, qval)

llfIE4_Q(x, nvec, dvec, yvec, qval)

llfH4_Q(x, nvec, dvec, yvec, qval)

llfLN4_Q(x, nvec, dvec, yvec, qval)

llfG4_Q(x, nvec, dvec, yvec, qval)

llfQE4_Q(x, nvec, dvec, yvec, qval)

llfP4_Q(x, nvec, dvec, yvec, qval)

llfL4_Q(x, nvec, dvec, yvec, qval)

llfSM_Q(x, nvec, dvec, yvec, qval)

llfH0_Q(x, nvec, dvec, yvec, qval)
}
\arguments{
\item{x}{parameter values}

\item{nvec}{vector containing the number of observations at each dose level}

\item{dvec}{vector containing the unique ordered dose levels}

\item{yvec}{vector containing the number of adverse events at each dose level}

\item{qval}{BMR}
}
\value{
loglikelihood value
}
\description{
Loglikelihood functions for the different dose response models, for quantal data
}
