% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_Q_Cov}
\alias{llfE4_Q_Cov}
\alias{llfIE4_Q_Cov}
\alias{llfH4_Q_Cov}
\alias{llfLN4_Q_Cov}
\alias{llfG4_Q_Cov}
\alias{llfQE4_Q_Cov}
\alias{llfP4_Q_Cov}
\alias{llfL4_Q_Cov}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfIE4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfH4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfLN4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfG4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfQE4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfP4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)

llfL4_Q_Cov(
  pars,
  x,
  n,
  y,
  qval,
  covar = c("background", "BMD_d", "all", "none"),
  nlevels,
  trt_ind
)
}
\arguments{
\item{pars}{parameter values}

\item{x}{vector containing the unique ordered dose levels}

\item{n}{vector containing the number of observations at each dose level}

\item{y}{vector containing the number of adverse events at each dose level}

\item{qval}{BMR}

\item{covar}{which parameter includes a covariate effect}

\item{nlevels}{number of covariate levels}

\item{trt_ind}{matrix indicating which reponse corresponds to which covariate level}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
