% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{LNtoN}
\alias{LNtoN}
\title{Function to convert geometric to arithmetic mean/standard deviation}
\usage{
LNtoN(gm, gsd)
}
\arguments{
\item{gm}{Geometric means per odered dose level, on original scale}

\item{gsd}{Geometric standard deviations per odered dose level, on original scale}
}
\value{
Vector containing the arithmetic means and standard deviations per ordered dose level.
}
\description{
Converts geometric mean (sd) on original scale to arithmetic mean (sd) to be used in normal distribution
}
\examples{
LNtoN(gm = immunotoxicityData$Mean[1:5], gsd = immunotoxicityData$SD[1:5])

}
