% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ModelTestC.R
\name{modelTestC}
\alias{modelTestC}
\title{function to perform MCMC sampling for all the dose-response model.}
\usage{
modelTestC(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)
}
\arguments{
\item{best.fit}{best fitting model}

\item{data.N}{list containing data values for the normal models}

\item{data.LN}{list containing data values for the lognormal models}

\item{stanBest}{stan object for the best fitting model}

\item{type}{estimation type. Laplace approximation or McMC}

\item{seed}{random seed for reproducibility. Defaults to 123}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{nrchains}{number of MCMC chains. Defaults to 3}

\item{nriterations}{number of MCMC iterations.Defaults to 3000}

\item{warmup}{number of MCMC iterations for warmup. Defaults to 1000}

\item{delta}{adapt_delta value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 0.8.}

\item{treedepth}{tree_depth value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 10.}
}
\description{
function to perform MCMC sampling for all the dose-response model.
}
\examples{
{


}


}
