% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modelTestQ.R
\name{modelTestQ}
\alias{modelTestQ}
\title{Test whether the best fit model fits equally well as the saturated ANOVA model}
\usage{
modelTestQ(
  best.fit,
  data.Q,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)
}
\arguments{
\item{best.fit}{name of the best fitting model}

\item{stanBest}{fit of the best fitting model as returned by rstan's optimizing or sampling routine}

\item{type}{either 'MCMC' or 'Laplace'}

\item{seed}{default 123}

\item{ndraws}{the number of draws from the posterior, default 30000}

\item{nrchains}{the number of chains to be used in the MCMC}

\item{nriterations}{the number of iterations per chain}

\item{warmup}{the number of iterations per chain to be discarded as burnin}

\item{delta}{default 0.8}

\item{treedepth}{default 10}

\item{data.N}{the input data as returned by function PREP_DATA_N}

\item{data.LN}{the input data as returned by function PREP_DATA_LN}
}
\value{
Bayes factor for the best fitting model compared to the saturated ANOVA model. A Bayes factor < 10 indicates the best fitting model fits equally well as the ANOVA model.
}
\description{
More detailed descriprion
}
