% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{NLN_test}
\alias{NLN_test}
\title{Performs Shapiro-Wilk's test for normality on individual dose-response data}
\usage{
NLN_test(datind)
}
\arguments{
\item{datind}{a dataframe containing dose (must be named x) and response named y}
}
\value{
a list of p-values and messages confirming the normality or lognormality of the data
}
\description{
Performs Shapiro-Wilk's test for normality on individual dose-response data
}
