% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{NtoLN}
\alias{NtoLN}
\title{Function to convert arithmetic to geometric mean/standard deviation}
\usage{
NtoLN(am, asd)
}
\arguments{
\item{am}{Arithmetic means per odered dose level, on original scale}

\item{asd}{Arithmetic standard deviations per odered dose level, on original scale}
}
\value{
Vector containing the geometric means and standard deviations per ordered dose level.
}
\description{
Converst arithmetic mean (sd) on original scale to geometric mean (sd) to be used in log-normal distribution
}
\examples{
 data("immunotoxicityData.rda")
 NtoLN(am = immunotoxicityData$Mean, asd = immunotoxicityData$SD)

}
