% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outLP}
\alias{outLP}
\title{Function for internal use}
\usage{
outLP(par_obj, pvec, max.dose)
}
\arguments{
\item{par_obj}{value}

\item{pvec}{value}

\item{max.dose}{logical}
}
\value{
.
}
\description{
Function for internal use
}
