% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outLPQ}
\alias{outLPQ}
\title{Function for internal use}
\usage{
outLPQ(par_obj, pvec, max.dose, rho = FALSE)
}
\arguments{
\item{par_obj}{value}

\item{pvec}{value}

\item{max.dose}{logical}

\item{rho}{logical}
}
\value{
.
}
\description{
Function for internal use
}
