% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{par_extractC}
\alias{par_extractC}
\title{function to extract the sampled values for the model parameters and the BMD (clustered continuous data)}
\usage{
par_extractC(
  mod_obj,
  pars = c(letters[1:4], "k", paste0("par", 1:6, ""), "invsigma2", "mu_inf", "mu_0",
    "rho_cluster"),
  model_name
)
}
\arguments{
\item{mod_obj}{model object either of class stanfit or stanfitOptim}

\item{pars}{parameters to be extracted.}

\item{model_name}{name of the model to be extracted}
}
\value{
dataframe containing model estimates
}
\description{
function to extract the sampled values for the model parameters and the BMD (clustered continuous data)
}
