% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputC.R
\name{par_med}
\alias{par_med}
\title{function to get the median of the needed parameters for predicted values. This is done over all
fitted models}
\usage{
par_med(mod.obj, type = c("continuous", "quantal"), clustered = F)
}
\arguments{
\item{mod_obj}{BMDBMA model object}
}
\value{
dataframe containing parameter estimates per model
}
\description{
function to get the median of the needed parameters for predicted values. This is done over all
fitted models
}
