% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputQ.R
\name{parq_extract}
\alias{parq_extract}
\title{function to extract the sampled values for the model parameters and the BMD}
\usage{
parq_extract(
  mod_obj,
  pars = c(letters[c(1, 2, 4)], "BMD", paste0("par", 1:3), "min_response"),
  model_name,
  rho = FALSE
)
}
\arguments{
\item{mod_obj}{model object either of class stanfit or stanfitOptim}

\item{pars}{parameters to be extracted.}

\item{model_name}{name of the model to be extracted}

\item{rho}{should be set to TRUE if data is clustered}
}
\value{
dataframe containing model estimates
}
\description{
function to extract the sampled values for the model parameters and the BMD
}
