% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{pert_dist}
\alias{pert_dist}
\title{Functions to get parameters for PERT prior}
\usage{
pert_dist(x, lb, ub, s1, s2)
}
\arguments{
\item{x}{value}

\item{lb}{value}

\item{ub}{value}

\item{s1}{value}

\item{s2}{value}
}
\value{
.
}
\description{
Functions to get parameters for PERT prior
}
