% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{predict.BMADRQ}
\alias{predict.BMADRQ}
\title{Function to get predicted values from the DRMS}
\usage{
predict.BMADRQ(
  mod.obj,
  dose,
  what = c("predicted", "resp_at_BMD"),
  model_averaged = FALSE,
  weight_type = c("BS", "LP")
)
}
\arguments{
\item{mod.obj}{is the result of pars_med}

\item{what}{prediction of interest}

\item{model_averaged}{option to compute model averaged prediction}

\item{weight_type}{type of weight to be used}

\item{type}{dose response model type}
}
\value{
.
}
\description{
Function to get predicted values from the DRMS
}
