% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DataN.R
\name{PREP_DATA_N}
\alias{PREP_DATA_N}
\title{Function to set data in the correct format}
\usage{
PREP_DATA_N(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04
)
}
\arguments{
\item{data}{dataframe with input data, order of columns should be: dose, response, sd, n.}

\item{sumstats}{logical. TRUE indicates summary data is provided while FALSE indicates individual-level data.}

\item{geom.stats}{logical. TRUE if geometric summary data is provided.}

\item{sd}{logical. TRUE indicates that standard deviation per dose is provided,
FALSE indicates that standard error per dose level is provided}

\item{q}{the specified BMR}

\item{bkg}{vector containing informative prior for the background.
It should be specified as minimum, most likely and maximum. Defaults to NULL.}

\item{maxy}{vector containing informative prior for the maximum response.
It should be specified as minimum, most likely and maximum. Defaults to NULL.}

\item{prior.BMD}{vector containing informative prior for the BMD.
It should be specified as minimum, most likely and maximum. Defaults to NULL.}

\item{shape.a}{shape parameter that determines the flatness of the Pert prior for the background.
Defaults to 4, which implies a peak at the most likely value.}

\item{shape.c}{shape parameter that determines the flatness of the Pert prior for c.
Defaults to 4, which implies a peak at the most likely value.}

\item{shape.BMD}{shape parameter that determines the flatness of the Pert prior for the BMD.
Defaults to 0.0001, which implies a flat prior.}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
This function also generates appropriate start values and uninformative priors for each model
Input should be given as arithmetic mean and standard deviation on the original scale
}
