% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_N_C}
\alias{PREP_DATA_N_C}
\alias{PREP_DATA_LN_C}
\title{Function to set data in the correct input format for model-averaging functions, for clustered continuous data}
\usage{
PREP_DATA_N_C(
  data,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = T
)

PREP_DATA_LN_C(
  data,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = T
)
}
\arguments{
\item{data}{a dataframe with individual-level input data, order of columns should be: dose, response, litter}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional, can be NA), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{maxy}{vector containing minimum, most likely (optional, can be NA), and maximum value for the response at dose infinity. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional, can be NA), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.c}{shape parameter for the modified PERT distribution on parameter c, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior}

\item{prior.d}{prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior), N05 (for a N(0.5,0.5) prior), or 'custom'}

\item{d.mean}{mean of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.std}{sd of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.trunc}{truncation (upper) of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}
}
\value{
\code{data} list containing the data and prior distributions

\code{start} list containing start values for the model parameters

\code{startQ} list containing start values for the quadratic exponential model

\code{shapiro.p} p-value of the Shapiro-Wilks test for normality

\code{bartlett.p} p-value of the Bartlett test for homoskedasticity

\code{shapiro.msg} message with result of test for normality

\code{bartlett.msg} message with result of test for homoskedasticity
}
\description{
This function takes a dataset as input and generates the data list, starting values, and prior distributions needed by the stan
models to be fitted. Shape parameters for the prior distribution can be any value, with 0.0001 resulting in a uniform prior and 4 resulting in a peaked prior.
For example, setting the shape parameter to 2 results in a slightly peaked prior distribution.
}
\examples{

data.input <- das5$data
data.input <- data.frame(dose = data.input$dose, response = data.input$foetalweight, litter = data.input$number)

# default priors
data_N <- PREP_DATA_N_C(data.input, q = 0.05, prior.d = 'N11')

}
