% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_QA}
\alias{PREP_DATA_QA}
\title{Function to set data in the correct format, for quantal data}
\usage{
PREP_DATA_QA(
  data,
  sumstats = TRUE,
  q,
  bkg = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.BMD = 1e-04,
  cluster = FALSE,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = TRUE,
  extended.value = 3
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, number of adverse events (increasing), n (or dose, adverse event yes/no in case of individual data)}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided. If individual-level data are provided, a litter indicator should be included instead of n (column 3)}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior}

\item{cluster}{logical variable to indicate if data is clustered (i.e. litter effect). TRUE = clustered data. Defaults to FALSE}

\item{prior.d}{prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior), N05 (for a N(0.5,0.5) prior), or 'custom'}

\item{d.mean}{mean of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.std}{sd of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.trunc}{truncation (upper) of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}

\item{extended.value}{value for the upper limit of BMD prior in extended range}
}
\value{
\code{data} list containing data and prior distributions

\code{start} list containing start values for the model parameters

\code{startQ} list containing start values for the quadratic exponential model#'
}
\description{
This function takes a dataset as input and generates the data list, starting values, and prior distributions needed by the stan
models to be fitted. Shape parameters for the prior distribution can be any value, with 0.0001 resulting in a uniform prior and 4 resulting in a peaked prior.
For example, setting the shape parameter to 2 results in a slightly peaked prior distribution.
}
\examples{
# default priors
dataQ = PREP_DATA_QA(data = data_quantal, q = 0.1, sumstats = T)

# litter effect
dataQ = PREP_DATA_QA(data_quantal_litter, cluster = T, q = 0.1)

}
