% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputQ.R
\name{print.BMADRQ}
\alias{print.BMADRQ}
\title{print function for BMABMDR objects}
\usage{
\method{print}{BMADRQ}(mod.obj)
}
\arguments{
\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs and weights per model
}
\description{
print function for BMABMDR objects
}
