% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_SamplingQ.R
\name{samplingQ_MA}
\alias{samplingQ_MA}
\title{function to fit the BMD model using MCMC}
\usage{
samplingQ_MA(
  data.Q,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  nrchains = 3,
  nriterations = 5000,
  warmup = 1000,
  delta = 0.8,
  treedepth = 10,
  seed = 123,
  pvec = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{data.Q}{list containing the data, prior parameters and the starting values}

\item{prior.weights}{a vector of prior weights for the models. 1 imply a model should be used, 0 otherwise.
Defaults to rep(1, 8).}

\item{ndraws}{number of draws from the . Defaults to 30000.}

\item{nrchains}{number of MCMC chains to run. Defaults to 3.}

\item{nriterations}{total number of MCMC iterations. Defaults to 5000.}

\item{warmup}{total number of MCMC iterations to be used as warmup. Defaults to 1000.}

\item{delta}{adapt value for the MCMC chain. See \code{\link[rstan]{sampling}} for more.}

\item{treedepth}{adapt value for the MCMC chain. See \code{\link[rstan]{sampling}} for more.}

\item{seed}{random seed for reproducibility}

\item{pvec}{probability vector to compute credible interval for the BMD. Defaults to c(0.05,0.5,0.95).}
}
\value{
list containing the following results
\enumerate{
\item E4_Q parameter estimates from the exponential model
\item IE4_Q parameter estimates from the inverse-exponential model
\item H4_Q parameter estimates from the Hill model
\item LN4_Q parameter estimates from the lognormal model
\item G4_Q parameter estimates from the gamma model
\item QE4_Q parameter estimates from the quadratic-exponential model
\item P4_Q parameter estimates from the probit model
\item L4_Q parameter estimates from the logit model
\item MA_laplace Laplace approximation model averaged BMD estimates using all the models
\item MA_bs Bridge sampling model averaged BMD estimates using all the models
\item MA_bs_conv Bridge sampling model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item weights_bs model weights computed using bridge sampling
\item convergence vector indicating model convergence or not. 1 = converged, 0 otherwise.
\item llQ vector of model likelihoods
\item bf Bayes factor comparing the best model against saturated ANOVA model
}
}
\description{
function to fit the BMD model using MCMC
}
\examples{

}
