% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.BMADRQ}
\alias{summary.BMADRQ}
\title{Function for internal use}
\usage{
summary.BMADRQ(mod.obj, conv = FALSE)
}
\arguments{
\item{mod.obj}{is the result of pars_med}

\item{conv}{}
}
\value{
.
}
\description{
Function for internal use
}
