% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{weights_extract}
\alias{weights_extract}
\title{Function for internal use}
\usage{
weights_extract(mod.obj, type = c("BS", "LP", "both"))
}
\arguments{
\item{mod.obj}{value}

\item{type}{value}
}
\value{
.
}
\description{
Function for internal use
}
