% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponse.R
\name{anydoseresponseLN}
\alias{anydoseresponseLN}
\title{Function to determine if there is a dose-response effect (Lognormal version)}
\usage{
anydoseresponseLN(dose.a, mean.a, sd.a, n.a)
}
\arguments{
\item{dose.a}{ordered dose levels}

\item{mean.a}{mean response}

\item{sd.a}{standard deviation}

\item{n.a}{number of observations per dose level}
}
\value{
list containing Baye's factor and decision.
}
\description{
Function to determine if there is a dose-response effect (Lognormal version)
}
