% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponse.R
\name{anydoseresponseN}
\alias{anydoseresponseN}
\title{Function to determine if there is a dose-response effect, for the normal distribution}
\usage{
anydoseresponseN(dose.a, mean.a, sd.a, n.a)
}
\arguments{
\item{dose.a}{ordered dose levels}

\item{mean.a}{mean response}

\item{sd.a}{standard deviation}

\item{n.a}{number of observations per dose level}
}
\value{
list containing Bayes factor and decision.
}
\description{
This function tests for any dose-response effect using Bayes factor.
It fits a null model and a saturated model and compares these two models using model posterior probabilities obtained via bridge rstan::sampling.
}
\examples{

# We use the first 5 rows because those are observations from subjects belonging to the same group.
data("immunotoxicityData.rda")  #load the immunotoxicity data
anydoseresponseN(dose.a = immunotoxicityData$Dose[1:5],
                 mean.a = immunotoxicityData$Mean[1:5],
                 sd.a = immunotoxicityData$SD[1:5],
                 n.a = immunotoxicityData$n[1:5])

}
