% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponseQ.R
\name{anydoseresponseQ}
\alias{anydoseresponseQ}
\title{Function to determine if there is a dose-response effect}
\usage{
anydoseresponseQ(dose.a, y.a, n.a)
}
\arguments{
\item{dose.a}{dose levels}

\item{y.a}{number of adverse events per dose level}

\item{n.a}{number of observations per dose level}
}
\value{
text
}
\description{
Function to determine if there is a dose-response effect
}
