% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anydoseresponse.R
\name{anydoseresponseQ}
\alias{anydoseresponseQ}
\title{Function to determine if there is a dose-response effect, for quantal endpoints}
\usage{
anydoseresponseQ(dose.a, y.a, n.a, cluster = FALSE)
}
\arguments{
\item{dose.a}{ordered dose levels}

\item{y.a}{number of adverse events}

\item{n.a}{number of observations per dose level}

\item{cluster}{logical variable indicating whether the data are clustered or not}
}
\value{
list containing estimated Bayes factor and the decision.
}
\description{
This function tests for any dose-response effect using Bayes factor.
It fits a null model and a saturated model and compares these two models using model posterior probabilities obtained via bridge sampling.
}
