% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{bartlett}
\alias{bartlett}
\title{Bartlett function for testing constant variance and co-efficient of variation; H0 equal variances}
\usage{
bartlett(sd, n)
}
\arguments{
\item{sd}{standard deviation per dose group}

\item{n}{number of observations per dose group}
}
\value{
test statistic and p-value for the Bartlett's test
}
\description{
Bartlett function for testing constant variance and co-efficient of variation; H0 equal variances
}
\examples{
bartlett(immunotoxicityData$SD[1:5], immunotoxicityData$n[1:5])

}
