% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_fit.R
\name{basic.plot}
\alias{basic.plot}
\alias{basic.plotQ}
\title{Basic plotting for the methods using covariates}
\usage{
basic.plot(x, model_name, increasing)

basic.plotQ(x)
}
\arguments{
\item{x}{BMDBMA model object}

\item{model_name}{model for which the fit should be plotted (one of get_models('continuous'))}

\item{increasing}{logical indicating whether the continuous data are increasing (T) or decreasing (F)}
}
\value{
a basic plot
}
\description{
Basic plotting for the methods using covariates
}
