% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMABMDR-package.R
\docType{package}
\name{BMABMDR-package}
\alias{BMABMDR-package}
\alias{BMABMDR}
\title{The 'BMABMDR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.3. https://mc-stan.org
}
