% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDLU}
\alias{BMDLU}
\title{Function for internal use}
\usage{
BMDLU(mod.obj, pvec = c(0.05, 0.5, 0.95), type = c("continuous", "quantal"))
}
\arguments{
\item{mod.obj}{value}

\item{pvec}{value}

\item{type}{value}
}
\value{
.
}
\description{
Function for internal use
}
