% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDMA_extract}
\alias{BMDMA_extract}
\title{function to extract model averaged BMD}
\usage{
BMDMA_extract(mod.obj, conv = FALSE)
}
\arguments{
\item{conv}{logical to indicate if mixture should be only based on converged
models or all the models regardless of the convergence status}

\item{mod_obj}{BMDBMA model object}
}
\value{
dataframe containing BMDL, BMD and BMDU
}
\description{
function to extract model averaged BMD
}
