% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputQ.R
\name{BMDQmixture_extract}
\alias{BMDQmixture_extract}
\title{function to extract BMD mixture}
\usage{
BMDQmixture_extract(mod.obj, conv = FALSE)
}
\arguments{
\item{conv}{logical to indicate if mixture should be only based on converged
models or all the models regardless of the convergence status}

\item{mod_obj}{BMDBMA model object}
}
\value{
dataframe containing Mixture per model
}
\description{
function to extract BMD mixture
}
