% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDWeights}
\alias{BMDWeights}
\title{Function for internal use}
\usage{
BMDWeights(mod.obj, type = c("continuous", "quantal"))
}
\arguments{
\item{mod.obj}{is the result of pars_med}

\item{type}{value}
}
\value{
.
}
\description{
Function for internal use
}
