% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{BMDWeights}
\alias{BMDWeights}
\title{function to get BMDs and weights}
\usage{
BMDWeights(mod.obj, type = c("continuous", "quantal"))
}
\arguments{
\item{type}{type of endpoint, either 'continuous' or 'quantal'}

\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs and weights per model
}
\description{
function to get BMDs and weights
}
