% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputC.R, R/outputQ.R
\name{BMDWeights}
\alias{BMDWeights}
\title{function to get BMDs and weights}
\usage{
BMDWeights(mod.obj, type = c("continuous", "quantal"))

BMDWeights(mod.obj, type = c("continuous", "quantal"))
}
\arguments{
\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs and weights per model

datatframe containing BMDs and weights per model
}
\description{
function to get BMDs and weights

function to get BMDs and weights
}
