% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM.R
\name{DRM.E4_NI}
\alias{DRM.E4_NI}
\alias{DRM.IE4_NI}
\alias{DRM.H4_NI}
\alias{DRM.LN4_NI}
\alias{DRM.G4_NI}
\alias{DRM.QE4_NI}
\alias{DRM.P4_NI}
\alias{DRM.L4_NI}
\alias{DRM.E4_LNI}
\alias{DRM.IE4_LNI}
\alias{DRM.H4_LNI}
\alias{DRM.LN4_LNI}
\alias{DRM.G4_LNI}
\alias{DRM.QE4_LNI}
\alias{DRM.P4_LNI}
\alias{DRM.L4_LNI}
\alias{DRM.E4_ND}
\alias{DRM.IE4_ND}
\alias{DRM.H4_ND}
\alias{DRM.LN4_ND}
\alias{DRM.G4_ND}
\alias{DRM.QE4_ND}
\alias{DRM.P4_ND}
\alias{DRM.L4_ND}
\alias{DRM.E4_LND}
\alias{DRM.IE4_LND}
\alias{DRM.H4_LND}
\alias{DRM.LN4_LND}
\alias{DRM.G4_LND}
\alias{DRM.QE4_LND}
\alias{DRM.P4_LND}
\alias{DRM.L4_LND}
\title{The dose response models}
\usage{
DRM.E4_NI(par, x, q, shift = 0)

DRM.IE4_NI(par, x, q, shift = 0)

DRM.H4_NI(par, x, q, shift = 0)

DRM.LN4_NI(par, x, q, shift = 0)

DRM.G4_NI(par, x, q, shift = 0)

DRM.QE4_NI(par, x, q, shift = 0)

DRM.P4_NI(par, x, q, shift = 0)

DRM.L4_NI(par, x, q, shift = 0)

DRM.E4_LNI(par, x, q, shift)

DRM.IE4_LNI(par, x, q, shift)

DRM.H4_LNI(par, x, q, shift)

DRM.LN4_LNI(par, x, q, shift)

DRM.G4_LNI(par, x, q, shift)

DRM.QE4_LNI(par, x, q, shift)

DRM.P4_LNI(par, x, q, shift)

DRM.L4_LNI(par, x, q, shift)

DRM.E4_ND(par, x, q, shift = 0)

DRM.IE4_ND(par, x, q, shift = 0)

DRM.H4_ND(par, x, q, shift = 0)

DRM.LN4_ND(par, x, q, shift = 0)

DRM.G4_ND(par, x, q, shift = 0)

DRM.QE4_ND(par, x, q, shift = 0)

DRM.P4_ND(par, x, q, shift = 0)

DRM.L4_ND(par, x, q, shift = 0)

DRM.E4_LND(par, x, q, shift)

DRM.IE4_LND(par, x, q, shift)

DRM.H4_LND(par, x, q, shift)

DRM.LN4_LND(par, x, q, shift)

DRM.G4_LND(par, x, q, shift)

DRM.QE4_LND(par, x, q, shift)

DRM.P4_LND(par, x, q, shift)

DRM.L4_LND(par, x, q, shift)
}
\arguments{
\item{par}{parameters in the order a, c, d, b}

\item{x}{unique ordered dose levels}

\item{q}{specified BMR}
}
\value{
Vector containing the expected response at each dose level
}
\description{
the mean/median dose-response model per dose-level.
}
\details{
These are the dose response models used internally in the BMA functions.
}
\examples{
data("immunotoxicityData.rda")  #load the immunotoxicity data
DRM.E4_NI(par = c(1.06, 0.015, 0.2, 1), x = immunotoxicityData$Dose[1:5], q = 0.1)

}
