% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{flat}
\alias{flat}
\title{Function for internal use}
\usage{
flat(dose, mean, inc)
}
\arguments{
\item{dose}{value}

\item{mean}{value}

\item{inc}{logical}
}
\value{
.
}
\description{
Function for internal use
}
