% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNs.R
\name{flat}
\alias{flat}
\title{Function for internal use}
\usage{
flat(dose, mean, n, inc)
}
\arguments{
\item{dose}{value}

\item{mean}{value}

\item{n}{value}

\item{inc}{logical variable to indicate if the dose-resonse curve is increasing or decreasing}
}
\value{
.logical value indicating if the dose-response curve is flat or not
}
\description{
Function for internal use
}
