% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Laplace.R
\name{full.laplace_MA}
\alias{full.laplace_MA}
\title{Perform model averaging using Full Laplace method}
\usage{
full.laplace_MA(
  data.N,
  data.LN,
  prior.weights,
  priordist = "PERT",
  prior.BMD = FALSE,
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95),
  plot = FALSE
)
}
\arguments{
\item{data.N}{the input data as returned by function PREP_DATA_N}

\item{data.LN}{the input data as returned by function PREP_DATA_LN}

\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{priordist}{which prior distribution should be used (currently only "PERT" is implemented)}

\item{prior.BMD}{logical indicating whether an informative prior for the BMD is used (currently not implemented)}

\item{ndraws}{the number of draws, default 30000}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}

\item{plot}{logical indicating whether a simple plot of model fits should be shown}
}
\value{
List with the model-specific results, model weights, the model averaged BMD, BMDL and BMDU. In addition for each model a matrix with covariance/correlation between parameter b/BMD and parameter d is given. Some additional output used in other external functions is also given.
}
\description{
This method assumed data for continuous endpoints.
}
\details{
More detailed descriprion
}
