% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Laplace.R
\name{full.laplace_MA_Cov}
\alias{full.laplace_MA_Cov}
\title{Perform model averaging using Full Laplace method}
\usage{
full.laplace_MA_Cov(
  data,
  sumstats = TRUE,
  sd = TRUE,
  q = q,
  prior.d = prior.d,
  extended = TRUE,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data}{the summary data, with columns: dose, response, sd, n, covariate}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{sd}{logical indicating whether standard deviation (T, default) or standard error (F) is provided}

\item{q}{specified BMR}

\item{prior.d}{prior distribution for parameter d, should be either N11 (default) or EPA}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)#}

\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws, default 30000}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}
}
\value{
\enumerate{
\item E4_N parameter estimates from the (Exponential, Normal) model
\item IE4_N parameter estimates from the (Inverse Exponential, Normal) model
\item H4_N parameter estimates from the (Hill, Normal) model
\item LN4_N parameter estimates from the (Log-normal, Normal)
\item G4_N parameter estimates from the (Gamma, Normal) model
\item QE4_N parameter estimates from the (Quadratic Exponential, Normal) model
\item P4_N parameter estimates from the (Probit, Normal) model
\item L4_N parameter estimates from the (Logit, Normal) model
\item E4_LN parameter estimates from the (Exponential, Log-normal) model
\item IE4_LN parameter estimates from the (Inverse Exponential, Log-normal) model
\item H4_LN parameter estimates from the (Hill, Log-normal) model
\item LN4_LN parameter estimates from the (Log-normal, Log-normal)
\item G4_LN parameter estimates from the (Gamma, Log-normal) model
\item QE4_LN parameter estimates from the (Quadratic Exponential, Log-normal) model
\item P4_LN parameter estimates from the (Probit, Log-normal) model
\item L4_LN parameter estimates from the (Logit, Log-normal) model
\item MA  Model averaged BMD credible interval
\item weights Model weights used in the averaging
}
}
\description{
Using Laplace approximation, we compute the parameters and weights of each model.
}
