% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_LaplaceQ.R
\name{full.laplaceQ_MA}
\alias{full.laplaceQ_MA}
\title{function to fit the BMD model using Laplace approximation}
\usage{
full.laplaceQ_MA(
  data.Q,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data.Q}{list containing data values as given by \link{PREP_DATA_QA}}

\item{prior.weights}{model weights determining if a model is to be included in the model averaging or not.
1 implies model is included, 0 otherwise. Defaults to rep(1,8).}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{seed}{random seed for reproducibility. Defaults to 123}

\item{pvec}{probability vector to compute credible interval for the BMD. Defaults to c(0.05,0.5,0.95).}
}
\value{
list containing the following results
\enumerate{
\item E4_Q parameter estimates from the exponential model
\item IE4_Q parameter estimates from the inverse-exponential model
\item H4_Q parameter estimates from the Hill model
\item LN4_Q parameter estimates from the lognormal
\item G4_Q parameter estimates from the gamma model
\item QE4_Q parameter estimates from the quadratic-exponential model
\item P4_Q parameter estimates from the probit model
\item L4_Q parameter estimates from the logit model
\item MA_laplace Laplace approximation model averaged BMD estimates using all the models
\item MA_lp_conv Laplace approximation model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item convergence vector indicating model convergence or not. 1 = converged, 0 otherwise.
\item llQ vector of model likelihoods for normal distribution
\item bf Bayes factor comparing the best model against saturated ANOVA model
}
}
\description{
function to fit the BMD model using Laplace approximation
}
\examples{

}
