% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_optimQ.R
\name{fun_optimQ}
\alias{fun_optimQ}
\title{function to perform Laplace approximation for each of the dose-response model.}
\usage{
fun_optimQ(
  mod,
  data,
  stv,
  ndraws = ndraws,
  seed = seed,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{mod}{stan model file}

\item{data}{list containing data values to be passed to the stan model file}

\item{stv}{list of starting values}

\item{ndraws}{number of draws to be made from the posterior distribution}

\item{seed}{random seed for reproducibility}

\item{pvec}{probability vector to compute credible interval for the BMD}
}
\value{
a stan results object
}
\description{
function to perform Laplace approximation for each of the dose-response model.
}
\examples{

}
