% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE42_Q}
\alias{llfE42_Q}
\alias{llfIE42_Q}
\alias{llfH42_Q}
\alias{llfLN42_Q}
\alias{llfG42_Q}
\alias{llfQE42_Q}
\alias{llfP42_Q}
\alias{llfL42_Q}
\alias{llfSM2_Q}
\alias{llfH02_Q}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE42_Q(x, nvec, dvec, yvec, qval, rho)

llfIE42_Q(x, nvec, dvec, yvec, qval, rho)

llfH42_Q(x, nvec, dvec, yvec, qval, rho)

llfLN42_Q(x, nvec, dvec, yvec, qval, rho)

llfG42_Q(x, nvec, dvec, yvec, qval, rho)

llfQE42_Q(x, nvec, dvec, yvec, qval, rho)

llfP42_Q(x, nvec, dvec, yvec, qval, rho)

llfL42_Q(x, nvec, dvec, yvec, qval, rho)

llfSM2_Q(x, nvec, dvec, yvec, qval, rho)

llfH02_Q(x, nvec, dvec, yvec, qval, rho)
}
\arguments{
\item{x}{parameter values}

\item{nvec}{vector containing the number of observations at each dose level}

\item{dvec}{vector containing the unique ordered dose levels}

\item{yvec}{vector containing the number of adverse events at each dose level}

\item{qval}{BMR}

\item{rho}{intra-cluster correlation parameter}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
