% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE42_Q}
\alias{llfE42_Q}
\alias{llfIE42_Q}
\alias{llfH42_Q}
\alias{llfLN42_Q}
\alias{llfG42_Q}
\alias{llfQE42_Q}
\alias{llfL42_Q}
\title{Loglikelihood functions for the different dose response models (Beta-Binomial)}
\usage{
llfE42_Q(x, nvec, dvec, yvec, qval, rho)

llfIE42_Q(x, nvec, dvec, yvec, qval, rho)

llfH42_Q(x, nvec, dvec, yvec, qval, rho)

llfLN42_Q(x, nvec, dvec, yvec, qval, rho)

llfG42_Q(x, nvec, dvec, yvec, qval, rho)

llfQE42_Q(x, nvec, dvec, yvec, qval, rho)

llfL42_Q(x, nvec, dvec, yvec, qval, rho)
}
\arguments{
\item{x}{parameter values}

\item{nvec}{number of observations}

\item{dvec}{dose levels}

\item{yvec}{response}

\item{qval}{BMR}

\item{rho}{intra-cluster correlation parameter}
}
\value{
loglikelihood value at x.
}
\description{
Loglikelihood functions for the different dose response models (Beta-Binomial)
}
