% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{llfE4_NI}
\alias{llfE4_NI}
\alias{llfIE4_NI}
\alias{llfH4_NI}
\alias{llfLN4_NI}
\alias{llfG4_NI}
\alias{llfQE4_NI}
\alias{llfP4_NI}
\alias{llfSM_N}
\alias{llfL4_NI}
\alias{llfE4_LNI}
\alias{llfIE4_LNI}
\alias{llfH4_LNI}
\alias{llfLN4_LNI}
\alias{llfG4_LNI}
\alias{llfQE4_LNI}
\alias{llfP4_LNI}
\alias{llfL4_LNI}
\alias{llfSM_LN}
\alias{llfE4_ND}
\alias{llfIE4_ND}
\alias{llfH4_ND}
\alias{llfLN4_ND}
\alias{llfG4_ND}
\alias{llfQE4_ND}
\alias{llfP4_ND}
\alias{llfL4_ND}
\alias{llfE4_LND}
\alias{llfIE4_LND}
\alias{llfH4_LND}
\alias{llfLN4_LND}
\alias{llfG4_LND}
\alias{llfQE4_LND}
\alias{llfP4_LND}
\alias{llfL4_LND}
\alias{llfE4_LNIc}
\title{Loglikelihood functions for the different dose response models}
\usage{
llfE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfH4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfG4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfP4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfSM_N(x, nvec, dvec, mvec, s2vec)

llfL4_NI(x, nvec, dvec, mvec, s2vec, qval)

llfE4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfIE4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfH4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfLN4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfG4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfQE4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfP4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfL4_LNI(x, nvec, dvec, mvec, s2vec, qval, shift)

llfSM_LN(x, nvec, dvec, mvec, s2vec, shift)

llfE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfIE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfH4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfLN4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfG4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfQE4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfP4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfL4_ND(x, nvec, dvec, mvec, s2vec, qval)

llfE4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfIE4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfH4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfLN4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfG4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfQE4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfP4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfL4_LND(x, nvec, dvec, mvec, s2vec, qval, shift)

llfE4_LNIc(x, d, n, nij, y, qval, shift)
}
\arguments{
\item{x}{parameter values}

\item{nvec}{vector containing the number of observations at each dose level}

\item{dvec}{unique ordered dose levels}

\item{mvec}{vector containing the response at each dose level}

\item{s2vec}{variance}

\item{qval}{BMR}

\item{shift}{value of the shift for negative geometric means}
}
\value{
.
}
\description{
Loglikelihood functions for the different dose response models
}
\examples{

}
