% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Conversion.R
\name{LNtoN}
\alias{LNtoN}
\title{Function to convert geometric to arithmetic mean/standard deviation}
\usage{
LNtoN(gm, gsd)
}
\arguments{
\item{gm}{Geometric means per odered dose level, on original scale}

\item{gsd}{Geometric standard deviations per odered dose level, on original scale}
}
\value{
Vector containing the arithmetic means and standard deviations per ordered dose level.
}
\description{
Converst arithmetic mean (sd) on original scale to geometric mean (sd) to be used in LNI method
}
\details{
Converst geometric mean (sd) to arithmetic mean (sd)
}
\examples{
 data("immunotoxicityData.rda")
 LNtoN(gm = immunotoxicityData$Mean, gsd = immunotoxicityData$SD)

}
