% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modelTest.R
\name{modelTest}
\alias{modelTest}
\alias{modelTestC}
\alias{modelTestQ}
\title{Function used internally to test whether the best fitting model fits the data well, compared to a saturated (non-monotone) model}
\usage{
modelTest(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)

modelTestC(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)

modelTestQ(
  best.fit,
  data.Q,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)
}
\arguments{
\item{best.fit}{model name}

\item{data.N}{object as given by PREP_DATA_N}

\item{data.LN}{object as given by PREP_DATA_LN}

\item{stanBest}{stan object for the model given in \code{best.fit}}

\item{type}{estimation type. Laplace or MCMC}

\item{seed}{random seed for reproducibility}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{nrchains}{number of MCMC chains. Defaults to 3}

\item{nriterations}{number of MCMC iterations.Defaults to 3000}

\item{warmup}{number of MCMC iterations for warmup. Defaults to 1000}

\item{delta}{adapt_delta value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 0.8.}

\item{treedepth}{tree_depth value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 10.}

\item{data.Q}{object as given by PREP_DATA_Q}
}
\value{
Bayes factor for the best fitting model compared to the saturated ANOVA model.
}
\description{
The function compares a dose-response model to the saturated model using an approximation to BIC.
}
