% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modelTest.R
\name{modelTest}
\alias{modelTest}
\title{function to perform MCMC sampling for all the dose-response model.}
\usage{
modelTest(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)
}
\arguments{
\item{best.fit}{best fitting model name}

\item{data.N}{list containing data values for the normal models}

\item{data.LN}{list containing data values for the lognormal models}

\item{stanBest}{stan object for the best fitting model}

\item{type}{estimation type. Laplace approximation or McMC}

\item{seed}{random seed for reproducibility. Defaults to 123}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{nrchains}{number of MCMC chains. Defaults to 3}

\item{nriterations}{number of MCMC iterations.Defaults to 3000}

\item{warmup}{number of MCMC iterations for warmup. Defaults to 1000}

\item{delta}{adapt_delta value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 0.8.}

\item{treedepth}{tree_depth value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 10.}
}
\description{
The function compares the best fitting model to the saturated model using the BIC.
}
\examples{

# we use the first 5 rows because those are observations from subjects belonging to the same group.
 data("immunotoxicityData.rda")  #load the immunotoxicity data
 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]),
                       sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors
 data_LN <- PREP_DATA_LN(data = as.data.frame(immunotoxicityData[1:5,]),
                         sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors

 pvec <- c(0.05, 0.5, 0.95)
 nrch=3;nriter=3000;wu=1000;dl=0.8;trd=10;sd=123;ndr=30000
 modelTest(best.fit = 'E4_N', data.N = data_N, data.LN = data_LN,
           stanBest = 'mE4_N', type = 'Laplace',
           seed = 123, ndraws= 30000, nrchains = 3,
           nriterations=nriter, warmup=wu, delta=dl, treedepth=trd)


}
