% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modelTest.R
\name{modelTestQ}
\alias{modelTestQ}
\title{Function to test whether the best fitting model fits the data well, compared to a saturated (non-monotone) model}
\usage{
modelTestQ(
  best.fit,
  data.Q,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)
}
\arguments{
\item{best.fit}{best fitting model name}

\item{data.Q}{object as given by PREP_DATA_QA}

\item{stanBest}{stan object for the best fitting model}

\item{type}{estimation type. Laplace or MCMC}

\item{seed}{random seed for reproducibility}

\item{ndraws}{number of draws to be made from the posterior distribution. Defaults to 30000}

\item{nrchains}{number of MCMC chains. Defaults to 3}

\item{nriterations}{number of MCMC iterations.Defaults to 3000}

\item{warmup}{number of MCMC iterations for warmup. Defaults to 1000}

\item{delta}{adapt_delta value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 0.8.}

\item{treedepth}{tree_depth value for the HMC in stan. See \code{\link[rstan]{sampling}} for more.
Defaults to 10.}
}
\value{
Bayes factor for the best fitting model compared to the saturated ANOVA model. A Bayes factor > 0.1 indicates the best fitting model fits equally well as the ANOVA model.
}
\description{
The function compares the best fitting model to the saturated model using an approximation to BIC.
}
\examples{

}
