% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outLP}
\alias{outLP}
\title{output function for estimated model parameter from each model}
\usage{
outLP(par_obj, pvec, max.dose)
}
\arguments{
\item{par_obj}{list containing parameter draws from either Laplace or MCMC fit}

\item{pvec}{probability vector to compute credible interval for the BMD.}

\item{max.dose}{maximum dose tested}
}
\value{
dataframe containing model estimates
}
\description{
output function for estimated model parameter from each model
}
