% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outLPQ}
\alias{outLPQ}
\title{Output function for estimated model parameter from each model (quantal data)}
\usage{
outLPQ(par_obj, pvec, max.dose, rho = FALSE)
}
\arguments{
\item{par_obj}{list containing parameter draws from either Laplace or MCMC fit}

\item{pvec}{probability vector to compute credible interval for the BMD.}

\item{max.dose}{maximum dose tested}

\item{rho}{should be set to TRUE if data is clustered}
}
\value{
dataframe containing model estimates
}
\description{
Output function for estimated model parameter from each model (quantal data)
}
