% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{par_extract}
\alias{par_extract}
\title{Function for internal use}
\usage{
par_extract(
  mod_obj,
  pars = c(letters[1:4], "BMD", paste0("par[", 1:5, "]"), "invsigma2", "min_response",
    "max_response"),
  model_name
)
}
\arguments{
\item{mod_obj}{value}

\item{pars}{value}
}
\value{
.
}
\description{
Function for internal use
}
