% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{par_med}
\alias{par_med}
\title{Function to get the median of the needed parameters for predicted values}
\usage{
par_med(mod.obj, type = c("continuous", "quantal"), clustered = F)
}
\arguments{
\item{type}{type of endpoint, either 'continuous' or 'quantal'}

\item{clustered}{logical indicating whether data are clustered (T) or not (F, default)}

\item{mod_obj}{BMDBMA model object}
}
\value{
dataframe containing parameter estimates per model
}
\description{
Function to get the median of the needed parameters for predicted values
}
