% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{par_med}
\alias{par_med}
\title{Function for internal use}
\usage{
par_med(mod.obj, type = c("continuous", "quantal"))
}
\arguments{
\item{mod.obj}{value}

\item{type}{value}
}
\value{
.
}
\description{
Function for internal use
}
