% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{parq_extract}
\alias{parq_extract}
\title{Function for internal use}
\usage{
parq_extract(
  mod_obj,
  pars = c(letters[c(1, 2, 4)], "BMD", paste0("par", 1:3), "min_response"),
  model_name,
  rho = FALSE
)
}
\arguments{
\item{mod_obj}{value}

\item{pars}{value}

\item{model_name}{model name}

\item{rho}{logical}
}
\value{
.
}
\description{
Function for internal use
}
