% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plot.BMADR}
\alias{plot.BMADR}
\title{Function for internal use}
\usage{
plot.BMADR(
  mod.obj,
  weight_type = c("BS", "LP"),
  include_data = TRUE,
  all = TRUE,
  title
)
}
\arguments{
\item{mod.obj}{fitted models}

\item{weight_type}{BS or LP}

\item{include_data}{include data points in plot}

\item{all}{all 4 plots in one plot}

\item{title}{title for plot if all = TRUE}
}
\value{
.
}
\description{
Function for internal use
}
