% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plot.BMADR}
\alias{plot.BMADR}
\title{Function to plot the results for continuous endpoints}
\usage{
plot.BMADR(
  mod.obj,
  type = c("increasing", "decreasing"),
  clustered = F,
  weight_type = c("BS", "LP"),
  include_data = TRUE,
  all = TRUE,
  title
)
}
\arguments{
\item{mod.obj}{BMDBMA model object}

\item{type}{dose-response type. It can either be "increasing" or "decreasing"}

\item{clustered}{logical indicating whether clustered data is used}

\item{weight_type}{type of model weight to be plotted. It can either be "BS" or "LP"}

\item{include_data}{logical argument to indicate if data should be included in the plots. Defaults to TRUE}

\item{all}{logical argument to indicate if all plots should be displayed}

\item{title}{title of the plot}
}
\value{
.
}
\description{
Function to plot the results for continuous endpoints
}
