% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plot.BMADRQ}
\alias{plot.BMADRQ}
\title{plot function for the BMD estimates, model weights, model predictions and model average predictions (Quantal dat)}
\usage{
plot.BMADRQ(
  mod.obj,
  weight_type = c("BS", "LP"),
  include_data = TRUE,
  all = TRUE,
  title
)
}
\arguments{
\item{mod.obj}{BMDBMA model object}

\item{weight_type}{type of model weight to be plotted. It can either be "BS" or "LP"}

\item{include_data}{logical argument to indicate if data should be included in the plots. Defaults to TRUE}

\item{all}{logical argument to indicate if all plots should be displayed}

\item{title}{title of the plot}
}
\value{
object of class ggplot
}
\description{
plot function for the BMD estimates, model weights, model predictions and model average predictions (Quantal dat)
}
