% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plot_BMD}
\alias{plot_BMD}
\title{Function for internal use}
\usage{
plot_BMD(
  BMD_DF,
  BMD.true = NULL,
  x,
  BMD,
  BMDL,
  BMDU,
  xlab = "Dataset",
  ylab = expression(log[10](BMD)),
  title = "BMD Estimate"
)
}
\arguments{
\item{BMD_DF}{dataframe containing BMD, BMDL and BMDU}

\item{BMD.true}{true value for BMD. Only necessary if such exist}

\item{x}{doses}

\item{BMD}{column name for BMD}

\item{BMDL}{column name for BMDL}

\item{BMDU}{column name for BMDU.}
}
\value{
.
}
\description{
Function for internal use
}
