% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Function to plot prior vs posterior distribution}
\usage{
plot_prior(mod.obj, data, model_name, parms = T)
}
\arguments{
\item{mod.obj}{the model object returned by full laplace or sampling method}

\item{data}{the data for a specific model used by the above method}

\item{model_name}{the model for which plots should be generated}

\item{parms}{if TRUE, model parameters (background, fold change, BMD, d) are shown; if FALSE, background, maximum response, BMD and d are shown}
}
\value{
.
}
\description{
Function to plot prior vs posterior distribution
}
