% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\alias{plot_priorQ}
\title{Plot function for the prior and posterior distribution of model parameters}
\usage{
plot_prior(mod.obj, data, model_name, parms = TRUE, clustered = FALSE)

plot_priorQ(mod.obj, data, model_name)
}
\arguments{
\item{mod.obj}{BMDBMA model object}

\item{data}{data list obtained using \code{\link{PREP_DATA_N}} or \code{\link{PREP_DATA_LN}}}

\item{model_name}{name of the model whose parameters are to be plotted (one of \code{get_models()})}

\item{parms}{logical indicating if fold change (T, default) or maximum response (F) should be plotted}

\item{clustered}{logical indicating whether the data are clustered (T) or not (F, default)}
}
\value{
object of class ggplot
}
\description{
Plot function for the prior and posterior distribution of model parameters
}
\examples{
data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE, sd = TRUE, q = 0.1)
data_LN <- PREP_DATA_LN(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE, sd = TRUE, q = 0.1)
FLBMD <- full.laplace_MA(data_N,data_LN)
plot_prior(FLBMD, data_N$data, 'E4_N')
plot_prior(FLBMD, data_LN$data, 'E4_LN')

}
