% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priorQ.R
\name{plot_priorQ}
\alias{plot_priorQ}
\title{Function to plot the prior and posterior distributions of the BMD and parameter estimates}
\usage{
plot_priorQ(mod.obj, data, model_name)
}
\arguments{
\item{mod.obj}{BMDBMA model object}

\item{data}{data list from the \code{\link{PREP_DATA_Q}}}

\item{model_name}{name of the model whose parameters are to be plotted. It can be eiter of
c("E4_Q","IE4_Q","H4_Q","LN4_Q","G4_Q","QE4_Q","P4_Q","L4_Q")}
}
\value{
object of class ggplot
}
\description{
Function to plot the prior and posterior distributions of the BMD and parameter estimates
}
