% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{posterior_diag}
\alias{posterior_diag}
\title{Function for internal use}
\usage{
posterior_diag(
  model_stan,
  pars = c("par1", "par2", "par3", "d", "invsigma2", "lp__")
)
}
\arguments{
\item{model_stan}{stan model}

\item{pars}{parameters}
}
\value{
.
}
\description{
Function for internal use
}
