% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{predict.BMADR}
\alias{predict.BMADR}
\title{Function for internal use}
\usage{
predict.BMADR(
  mod.obj,
  dose,
  type = c("increasing", "decreasing"),
  what = c("predicted", "resp_at_BMD"),
  model_averaged = FALSE,
  weight_type = c("BS", "LP")
)
}
\arguments{
\item{mod.obj}{is the result of pars_med}

\item{type}{dose response model type}

\item{what}{prediction of interest}

\item{model_averaged}{option to compute model averaged prediction}

\item{weight_type}{type of weight to be used}
}
\value{
.
}
\description{
Function for internal use
}
