% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{predict.BMADR}
\alias{predict.BMADR}
\title{function to get predicted values from the DRMS}
\usage{
predict.BMADR(
  mod.obj,
  dose,
  type = c("increasing", "decreasing"),
  what = c("predicted", "resp_at_BMD"),
  model_averaged = FALSE,
  weight_type = c("BS", "LP")
)
}
\arguments{
\item{mod.obj}{BMDBMA model object}

\item{dose}{vector of dose values to use for prediction}

\item{type}{dose response model type}

\item{what}{prediction of interest. It can either be "predicted" or "response_at_BMD"}

\item{model_averaged}{option to compute model averaged prediction. Defaults to FALSE}

\item{weight_type}{type of weight to be used. It can either be "BS" for bridge sampling or
"LP" for Laplace approximation}
}
\value{
dataframe of model prediction per model or
list with dataframe of model preditions and model-averaged predictions
}
\description{
function to get predicted values from the DRMS
}
