% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_N}
\alias{PREP_DATA_N}
\alias{PREP_DATA_LN}
\title{Function to set data in the correct input format for model-averaging functions.
This function also generates appropriate start values and uninformative priors for each model}
\usage{
PREP_DATA_N(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  extended = TRUE
)

PREP_DATA_LN(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  extended = TRUE
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, response, sd (or se), n, covariate (if given)}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{geom.stats}{logicial indicating whether, if summary data are provided, these are geometric (T) or arithmetic (F, default) summary statistics}

\item{sd}{logical indicating whether standard deviation (T, default) or standard error (F) is provided}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional, can be NA), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{maxy}{vector containing minimum, most likely (optional, can be NA), and maximum value for the response at dose infinity. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional, can be NA), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.c}{shape parameter for the modified PERT distribution on parameter c, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior}

\item{prior.d}{prior distribution for parameter d, should be either N11 (default) or EPA}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}
}
\value{
List with data and start values in correct format to be directly used within the model-averaging functions.
}
\description{
The function a dataset as input and generates the data list and starting values needed by the stan
models to be fitted.
}
\examples{

 # we use the first 5 rows because those are observations from subjects belonging to the same group.
 data("immunotoxicityData.rda")  #load the immunotoxicity data
 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]),
 sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors

 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE,
                       sd = TRUE, q = 0.1, bkg = c(0.62, 1.34, 2.06)) #example with informative prior on background


 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE,
                       sd = TRUE, q = 0.1,
                       prior.BMD = c(0.06, 0.25, 1)) #example with informative priors on the BMD


}
