% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DataN.R
\name{PREP_DATA_N}
\alias{PREP_DATA_N}
\title{Function to set data in the correct format}
\usage{
PREP_DATA_N(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, response, sd (or se), n}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{geom.stats}{logicial indicating whether, if summary data are provided, these are geometric (T) or arithmetic (F, default) summary statistics}

\item{sd}{logical indicating whether standard deviation (T, default) or standard error (F) is provided}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely, and maximum value for the background response}

\item{maxy}{vector containing minimum, most likely, and maximum value for the response at dose infinity}

\item{prior.BMD}{vector containing minimum, most likely, and maximum value for the BMD}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4, a value of 0.0001 implies a uniform distribution}

\item{shape.c}{shape parameter for the modified PERT distribution on parameter c, defaults to 4, a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 (implies a uniform distribution)}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
This function also generates appropriate start values and uninformative priors for each model
}
