% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_N}
\alias{PREP_DATA_N}
\alias{PREP_DATA_LN}
\title{Function to set data in the correct input format for model-averaging functions, for continuous data}
\usage{
PREP_DATA_N(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  extended = TRUE
)

PREP_DATA_LN(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  extended = TRUE
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, response, SD or SE (for summary data only), n (for summary data only), covariate level (if given)}

\item{sumstats}{logical indicating whether summary (TRUE, default) or individual-level (FALSE) data is provided}

\item{geom.stats}{logicial indicating whether, if summary data are provided, these are geometric (TRUE) or arithmetic (FALSE, default) summary statistics}

\item{sd}{logical indicating whether standard deviation (TRUE, default) or standard error (FALSE) is provided}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional, can be NA), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{maxy}{vector containing minimum, most likely (optional, can be NA), and maximum value for the response at dose infinity. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional, can be NA), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.c}{shape parameter for the modified PERT distribution on parameter c, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 for a peaked informative prior}

\item{prior.d}{prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior) or N05 (for a N(0.5,0.5) prior)}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}
}
\value{
\code{data} list containing data and prior distributions

\code{start} list containing start values for the model parameters

\code{startQ} list containing start values for the quadratic exponential model

\code{test.var} result of test for homoskedasticity

\code{test.NLN} result of normality test if individual-level data are given
}
\description{
This function takes a dataset as input and generates the data list, starting values, and prior distributions needed by the stan
models to be fitted. Shape parameters for the prior distribution can be any value, with 0.0001 resulting in a uniform prior and 4 resulting in a peaked prior.
For example, setting the shape parameter to 2 results in a slightly peaked prior distribution.
}
\examples{

# default priors
data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE, sd = TRUE, q = 0.1)

# informative prior on the background response
data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE, sd = TRUE, q = 0.1, bkg = c(0.62, 1.34, 2.06))

# informative prior on the BMD (peaked)
data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE,
                       sd = TRUE, q = 0.1,
                       prior.BMD = c(0.06, 0.25, 1), shape.BMD = 4)

# informative prior on the BMD (flat uniform)
data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]), sumstats = TRUE,
                       sd = TRUE, q = 0.1,
                       prior.BMD = c(0.06, 0.25, 1))

}
