% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_Q_COV}
\alias{PREP_DATA_Q_COV}
\title{Function to set data in the correct format, for quantal data with covariate effect (used internally)}
\usage{
PREP_DATA_Q_COV(
  data,
  sumstats = TRUE,
  q,
  bkg = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.BMD = 1e-04,
  cluster = FALSE,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = TRUE,
  extended.value = 3,
  covariate = "all"
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, number of adverse events, n, covariate level}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior}

\item{cluster}{logical variable to indicate if data is clustered. TRUE = clustered data. Defaults to FALSE}

\item{prior.d}{prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior) or N05 (for a N(0.5,0.5) prior)}

\item{d.mean}{mean of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.std}{sd of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{d.trunc}{truncation (upper) of the Normal prior for log(d), only used when prior.d = 'custom'}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}

\item{covariate}{for which parameter a covariate effect should be included. Defaults to 'all', other options are 'background' or 'BMD_d'}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
Function to set data in the correct format, for quantal data with covariate effect (used internally)
}
