% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DataQ.R
\name{PREP_DATA_QA}
\alias{PREP_DATA_QA}
\title{Function to set data in the correct format}
\usage{
PREP_DATA_QA(
  data,
  sumstats = TRUE,
  q,
  shape.a = 4,
  shape.BMD = 1e-04,
  cluster = FALSE,
  bkg = NULL,
  BMD_p = NULL
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, response, sd (or se), n}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided}

\item{q}{specified BMR}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4, a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 (implies a uniform distribution)}

\item{cluster}{logical indicating if data is clustered (defaults to FALSE)}

\item{bkg}{vector containing minimum, most likely, and maximum value for the background response}

\item{BMD_p}{vector containing minimum, most likely, and maximum value for the BMD}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
This function also generates appropriate start values and uninformative priors for each model
}
