% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DataQ.R
\name{PREP_DATA_QA}
\alias{PREP_DATA_QA}
\title{Function to set data in the correct format}
\usage{
PREP_DATA_QA(
  data,
  sumstats = TRUE,
  q,
  bkg = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.BMD = 1e-04,
  cluster = FALSE,
  prior.d = c("N11", "EPA"),
  extended = TRUE
)
}
\arguments{
\item{data}{dataframe with input data, order of columns should be: dose, response.}

\item{sumstats}{number of observations per dose level}

\item{q}{the specified BMR}

\item{bkg}{vector containing informative prior for the background.
It should be specified as minimum, most likely and maximum. Defaults to NULL.}

\item{prior.BMD}{vector containing informative prior for the BMD.
It should be specified as minimum, most likely and maximum. Defaults to NULL.}

\item{shape.a}{shape parameter that determines the flatness of the Pert prior for the background.
Defaults to 4, which implies a peak at the most likely value.}

\item{shape.BMD}{shape parameter that determines the flatness of the Pert prior for the BMD.
Defaults to 0.0001, which implies a flat prior.}

\item{cluster}{logical variable to indicate if data is clustered. TRUE = clustered data. Defaults to FALSE}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
This function also generates appropriate start values and uninformative priors for each model
Input should be given as arithmetic mean and standard deviation on the original scale
}
\examples{

}
