% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Data.R
\name{PREP_DATA_QA}
\alias{PREP_DATA_QA}
\title{Function to set data in the correct format}
\usage{
PREP_DATA_QA(
  data,
  sumstats = TRUE,
  q,
  bkg = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.BMD = 1e-04,
  cluster = FALSE,
  prior.d = "N11",
  extended = TRUE
)
}
\arguments{
\item{data}{a dataframe with input data, order of columns should be: dose, number of adverse events, n}

\item{sumstats}{logical indicating whether summary (T, default) or individual-level (F) data is provided. If individual-level data are provided, a litter indicator should be included instead of n (column 3)}

\item{q}{specified BMR}

\item{bkg}{vector containing minimum, most likely (optional), and maximum value for the background response. Defaults to NULL (non-informative prior)}

\item{prior.BMD}{vector containing minimum, most likely (optional), and maximum value for the BMD. Defaults to NULL (non-informative prior)}

\item{shape.a}{shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution}

\item{shape.BMD}{shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior}

\item{cluster}{logical variable to indicate if data is clustered. TRUE = clustered data. Defaults to FALSE}

\item{prior.d}{prior distribution for parameter d, should be either N11 (default) or EPA}

\item{extended}{logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)}
}
\value{
List with data and start values in correct format to be directly used within the BMA functions.
}
\description{
This function also generates appropriate start values and uninformative priors for each model
Input should be given as arithmetic mean and standard deviation on the original scale
}
