% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.BMADR}
\alias{print.BMADR}
\title{print function for BMABMDR objects}
\usage{
\method{print}{BMADR}(mod.obj, type = c("continuous", "quantal"))
}
\arguments{
\item{type}{type of endpoint, either 'continuous' or 'quantal'}

\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs and weights per model
}
\description{
print function for BMABMDR objects
}
