% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.BMADR}
\alias{print.BMADR}
\title{Function for internal use}
\usage{
\method{print}{BMADR}(mod.obj)
}
\arguments{
\item{mod.obj}{is the result of pars_med}
}
\value{
.
}
\description{
Function for internal use
}
