% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Sampling.R
\name{sampling_MA}
\alias{sampling_MA}
\alias{sampling_MAc}
\title{Perform model averaging using MCMC methods (Bridge sampling & Partial Laplace)}
\usage{
sampling_MA(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  nrchains = 3,
  nriterations = 3000,
  warmup = 1000,
  delta = 0.8,
  treedepth = 10,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)

sampling_MAc(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  nrchains = 3,
  nriterations = 3000,
  warmup = 1000,
  delta = 0.8,
  treedepth = 10,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{data.N}{the input data as returned by function PREP_DATA_N}

\item{data.LN}{the input data as returned by function PREP_DATA_LN}

\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws from the posterior, default 30000}

\item{nrchains}{the number of chains to be used in the MCMC}

\item{nriterations}{the number of iterations per chain}

\item{warmup}{the number of iterations per chain to be discarded as burnin}

\item{delta}{default 0.8}

\item{treedepth}{default 10}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}

\item{priordist}{which prior distribution should be used (currently only "PERT" is implemented)}

\item{prior.BMD}{logical indicating whether an informative prior for the BMD is used (currently not implemented)}
}
\value{
a list containing the following important entries:
\enumerate{
\item E4_N parameter estimates from the (Exponential, Normal) model
\item IE4_N parameter estimates from the (Inverse Exponential, Normal) model
\item H4_N parameter estimates from the (Hill, Normal) model
\item LN4_N parameter estimates from the (Log-normal, Normal)
\item G4_N parameter estimates from the (Gamma, Normal) model
\item QE4_N parameter estimates from the (Quadratic Exponential, Normal) model
\item P4_N parameter estimates from the (Probit, Normal) model
\item L4_N parameter estimates from the (Logit, Normal) model
\item E4_LN parameter estimates from the (Exponential, Log-normal) model
\item IE4_LN parameter estimates from the (Inverse Exponential, Log-normal) model
\item H4_LN parameter estimates from the (Hill, Log-normal) model
\item LN4_LN parameter estimates from the (Log-normal, Log-normal)
\item G4_LN parameter estimates from the (Gamma, Log-normal) model
\item QE4_LN parameter estimates from the (Quadratic Exponential, Log-normal) model
\item P4_LN parameter estimates from the (Probit, Log-normal) model
\item L4_LN parameter estimates from the (Logit, Log-normal) model
\item MA_laplace Partial Laplace model averaged BMD estimates using all included
\item MA_bridge_sampling Bridge sampling model averaged BMD estimates using all included models
\item MA_ls_conv Partial Laplace model averaged BMD estimates using only converged models
\item MA_bs_conv Bridge sampling model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item weights_bridge_sampling model weights computed using bridge sampling
\item convergence vector indicating model convergence or not. 1 = converged, 0 otherwise.
\item divergences vector indicating the proportion of divergent transitions for each model.
\item bf Bayes factor comparing the best model against saturated ANOVA model#'
}
}
\description{
Using MCMC, we compute the parameters of each model, perform model averaging using bridge sampling.
We also implemented a Laplace-MCMC method within this function where the model parameters are estimated
with MCMC, but the model weights are computed using Laplace approximation.
}
\details{
This method assumed data for continuous endpoints.

More detailed descriprion
}
\examples{
 # we use the first 5 rows because those are observations from subjects belonging to the same group.
 data("immunotoxicityData.rda")  #load the immunotoxicity data
 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]),
                       sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors
 data_LN <- PREP_DATA_LN(data = as.data.frame(immunotoxicityData[1:5,]),
                         sumstats = TRUE, sd = TRUE, q = 0.1) #example with default priors
 pvec <- c(0.05, 0.5, 0.95)
 prior.weights = rep(1, 16)
 nrch=3;nriter=3000;wu=1000;dl=0.8;trd=10;sd=123;ndr=30000
 SBMD = sampling_MA(data_N,data_LN,prior.weights,
                    ndraws=ndr,nrchains=nrch,
                    nriterations=nriter,warmup=wu,delta=dl,
                   treedepth=trd,seed=sd,pvec=pvec)


}
