% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Sampling.R
\name{samplingQ_MA}
\alias{samplingQ_MA}
\title{Perform model averaging using MCMC methods (Bridge sampling & Partial Laplace)}
\usage{
samplingQ_MA(
  data.Q,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  nrchains = 3,
  nriterations = 5000,
  warmup = 1000,
  delta = 0.8,
  treedepth = 10,
  seed = 123,
  pvec = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{prior.weights}{a vector specifying which of the 16 models should be included (1 = include, 0 = exclude)}

\item{ndraws}{the number of draws from the posterior, default 30000}

\item{nrchains}{the number of chains to be used in the MCMC}

\item{nriterations}{the number of iterations per chain}

\item{warmup}{the number of iterations per chain to be discarded as burnin}

\item{delta}{default 0.8}

\item{treedepth}{default 10}

\item{seed}{default 123}

\item{pvec}{vector specifying the three BMD quantiles of interest}

\item{data.N}{the input data as returned by function PREP_DATA_N}

\item{data.LN}{the input data as returned by function PREP_DATA_LN}

\item{priordist}{which prior distribution should be used (currently only "PERT" is implemented)}

\item{prior.BMD}{logical indicating whether an informative prior for the BMD is used (currently not implemented)}
}
\value{
a list containing the following important entries:
\enumerate{
\item E4_Q parameter estimates from the exponential model
\item IE4_Q parameter estimates from the inverse-exponential model
\item H4_Q parameter estimates from the Hill model
\item LN4_Q parameter estimates from the lognormal
\item G4_Q parameter estimates from the gamma model
\item QE4_Q parameter estimates from the quadratic-exponential model
\item P4_Q parameter estimates from the probit model
\item L4_Q parameter estimates from the logit model
\item MA_laplace Partial Laplace model averaged BMD estimates using all included
\item MA_bridge_sampling Bridge sampling model averaged BMD estimates using all included models
\item MA_ls_conv Partial Laplace model averaged BMD estimates using only converged models
\item MA_bs_conv Bridge sampling model averaged BMD estimates using only converged models
\item weights_laplace model weights using Laplace approximation to the posterior
\item weights_bridge_sampling model weights computed using bridge sampling
\item convergence vector indicating model convergence or not. 1 = converged, 0 otherwise.
\item divergences vector indicating the proportion of divergent transitions for each model.
\item bf Bayes factor comparing the best model against saturated ANOVA model#'
}
}
\description{
Using MCMC, we compute the parameters of each model, perform model averaging using bridge sampling.
We also implemented a Laplace-MCMC method within this function where the model parameters are estimated
with MCMC, but the model weights are computed using Laplace approximation.
}
\details{
This method assumed data for continuous endpoints.

More detailed descriprion
}
