% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputC.R
\name{summary.BMADR}
\alias{summary.BMADR}
\title{summary function for BMABMDR objects}
\usage{
summary.BMADR(mod.obj, conv = FALSE, clustered = FALSE)
}
\arguments{
\item{conv}{logical indicating if only converged models should be summarized.}

\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs, weights and parameters per model
}
\description{
summary function for BMABMDR objects
}
