% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.BMADR}
\alias{summary.BMADR}
\title{summary function for BMABMDR objects}
\usage{
summary.BMADR(
  mod.obj,
  type = c("continuous", "quantal"),
  conv = FALSE,
  clustered = FALSE
)
}
\arguments{
\item{type}{type of endpoint, can be 'continuous' or 'quantal'}

\item{conv}{logical indicating if only converged models should be summarized. Defaults to FALSE}

\item{clustered}{logical indicating if data are clustered (T) or not (F, default)}

\item{mod_obj}{BMDBMA model object}
}
\value{
datatframe containing BMDs, weights and parameters per model
}
\description{
summary function for BMABMDR objects
}
