% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{weights_extract}
\alias{weights_extract}
\title{function to turn the vector of weights into dataframe (continuous data)}
\usage{
weights_extract(mod.obj, type = c("BS", "LP", "both"))
}
\arguments{
\item{type}{type of weight to be extracted. It can be either of BS (bridge sampling), LP (Laplace) and
both to extract bridge sampling and laplace approximation.}

\item{mod_obj}{model object either of class stanfit or stanfitOptim}
}
\value{
dataframe containing model estimates
}
\description{
function to turn the vector of weights into dataframe (continuous data)
}
