% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputQ.R
\name{weightsQ_extract}
\alias{weightsQ_extract}
\title{function to extract the sampled values for the model parameters and the BMD}
\usage{
weightsQ_extract(mod.obj, type = c("BS", "LP", "both"))
}
\arguments{
\item{type}{type of weight to be extracted. It can be either of BS (bridge sampling), LP (Laplace) and
both to extract bridge sampling and laplace approximation.}

\item{mod_obj}{model object either of class stanfit or stanfitOptim}

\item{model_name}{name of the model to be extracted}
}
\value{
dataframe containing model estimates
}
\description{
function to extract the sampled values for the model parameters and the BMD
}
